@@ -1,52 +1,142 @@
 {
+  "name": "apollo-server-monorepo",
   "private": true,
   "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/apollographql/apollo-server.git"
-  },
+  "repository": "github:apollographql/apollo-server",
   "scripts": {
-    "compile": "lerna run compile",
-    "lint": "prettier-check --ignore-path .gitignore \"{docs/{,source/**},.,packages/**,test}/{*.js,*.ts}\"",
-    "lint-fix": "prettier --write --ignore-path .gitignore \"{docs/{,source/**},.,packages/**,test}/{*.js,*.ts}\"",
-    "prebootstrap": "npm install",
-    "postinstall": "lerna bootstrap",
-    "pretest": "npm run compile",
-    "test": "npm run testonly",
-    "posttest": "npm run lint",
-    "testonly": "lerna run test",
-    "coverage": "lerna run coverage",
-    "circle": "lerna run coverage && lerna run circle",
-    "release": "lerna publish --exact",
-    "precommit": "lint-staged"
-  },
-  "lint-staged": {
-    "*.ts": [
-      "prettier --write",
-      "git add"
-    ],
-    "*.js": [
-      "prettier --write",
-      "git add"
-    ]
+    "clean": "git clean -dfqX -- ./node_modules **/{dist,node_modules}/ ./packages/*/tsconfig*tsbuildinfo",
+    "compile": "tsc --build tsconfig.build.json",
+    "compile:clean": "tsc --build tsconfig.build.json --clean",
+    "watch": "tsc --build tsconfig.build.json --watch",
+    "release:version-bump": "lerna version -m 'Release' --ignore-changes='**/*.md'",
+    "release:version-bump:server": "npm run release:version-bump -- --force-publish=apollo-server,apollo-server-core,apollo-server-azure-functions,apollo-server-cloud-functions,apollo-server-cloudflare,apollo-server-express,apollo-server-fastify,apollo-server-hapi,apollo-server-koa,apollo-server-lambda,apollo-server-micro,apollo-server-integration-testsuite,apollo-server-testing",
+    "release:version-bump:federation": "npm run release:version-bump -- --force-publish=@apollo/federation,@apollo/gateway,apollo-federation-integration-testsuite",
+    "release:version-bump:server-and-federation": "npm run release:version-bump -- --force-publish=@apollo/federation,@apollo/gateway,apollo-federation-integration-testsuite,apollo-server,apollo-server-core,apollo-server-azure-functions,apollo-server-cloud-functions,apollo-server-cloudflare,apollo-server-express,apollo-server-fastify,apollo-server-hapi,apollo-server-koa,apollo-server-lambda,apollo-server-micro,apollo-server-integration-testsuite,apollo-server-testing",
+    "release:start-ci-publish": "node -p '`Publish (dist-tag:${process.env.APOLLO_DIST_TAG || \"latest\"})`' | git tag -F - \"publish/$(date -u '+%Y%m%d%H%M%S')\" && git push origin \"$(git describe --match='publish/*' --tags --exact-match HEAD)\"",
+    "postinstall": "lerna run prepare && npm run compile",
+    "test": "jest --verbose",
+    "test:clean": "jest --clearCache",
+    "test:watch": "jest --verbose --watchAll",
+    "testonly": "npm test",
+    "test:ci": "npm run coverage -- --ci --maxWorkers=2  --reporters=default --reporters=jest-junit",
+    "coverage": "npm test -- --coverage",
+    "coverage:upload": "codecov"
   },
   "engines": {
     "node": ">=6"
   },
+  "dependencies": {
+    "@apollographql/apollo-tools": "0.4.8",
+    "@apollo/federation": "file:packages/apollo-federation",
+    "@apollo/gateway": "file:packages/apollo-gateway",
+    "apollo-cache-control": "file:packages/apollo-cache-control",
+    "apollo-datasource": "file:packages/apollo-datasource",
+    "apollo-datasource-rest": "file:packages/apollo-datasource-rest",
+    "apollo-engine-reporting": "file:packages/apollo-engine-reporting",
+    "apollo-engine-reporting-protobuf": "file:packages/apollo-engine-reporting-protobuf",
+    "apollo-federation-integration-testsuite": "file:packages/apollo-federation-integration-testsuite",
+    "apollo-server": "file:packages/apollo-server",
+    "apollo-server-azure-functions": "file:packages/apollo-server-azure-functions",
+    "apollo-server-cache-memcached": "file:packages/apollo-server-cache-memcached",
+    "apollo-server-cache-redis": "file:packages/apollo-server-cache-redis",
+    "apollo-server-caching": "file:packages/apollo-server-caching",
+    "apollo-server-cloud-functions": "file:packages/apollo-server-cloud-functions",
+    "apollo-server-cloudflare": "file:packages/apollo-server-cloudflare",
+    "apollo-server-core": "file:packages/apollo-server-core",
+    "apollo-server-env": "file:packages/apollo-server-env",
+    "apollo-server-errors": "file:packages/apollo-server-errors",
+    "apollo-server-express": "file:packages/apollo-server-express",
+    "apollo-server-fastify": "file:packages/apollo-server-fastify",
+    "apollo-server-hapi": "file:packages/apollo-server-hapi",
+    "apollo-server-integration-testsuite": "file:packages/apollo-server-integration-testsuite",
+    "apollo-server-koa": "file:packages/apollo-server-koa",
+    "apollo-server-lambda": "file:packages/apollo-server-lambda",
+    "apollo-server-micro": "file:packages/apollo-server-micro",
+    "apollo-server-plugin-base": "file:packages/apollo-server-plugin-base",
+    "apollo-server-plugin-operation-registry": "file:packages/apollo-server-plugin-operation-registry",
+    "apollo-server-plugin-response-cache": "file:packages/apollo-server-plugin-response-cache",
+    "apollo-server-testing": "file:packages/apollo-server-testing",
+    "apollo-server-types": "file:packages/apollo-server-types",
+    "apollo-tracing": "file:packages/apollo-tracing",
+    "graphql-extensions": "file:packages/graphql-extensions"
+  },
   "devDependencies": {
-    "@types/graphql": "0.13.3",
-    "@types/jest": "23.3.1",
-    "@types/node": "10.5.5",
-    "graphql": "0.13.2",
-    "husky": "0.14.3",
-    "jest": "23.4.2",
-    "jest-matcher-utils": "23.2.0",
-    "lerna": "2.11.0",
-    "lint-staged": "7.2.0",
-    "prettier": "1.14.0",
-    "prettier-check": "2.0.0",
-    "supertest": "3.1.0",
-    "ts-jest": "23.1.2",
-    "typescript": "2.9.2"
+    "@types/async-retry": "1.4.2",
+    "@types/aws-lambda": "8.10.59",
+    "@types/body-parser": "1.19.0",
+    "@types/bunyan": "1.8.6",
+    "@types/connect": "3.4.33",
+    "@types/fast-json-stable-stringify": "2.0.0",
+    "@types/hapi": "17.8.8",
+    "@types/ioredis": "4.17.2",
+    "@types/jest": "25.2.3",
+    "@types/koa-router": "7.4.1",
+    "@types/lodash": "4.14.157",
+    "@types/lodash.sortby": "4.7.6",
+    "@types/lodash.xorby": "4.7.6",
+    "@types/loglevel": "1.5.4",
+    "@types/lru-cache": "5.1.0",
+    "@types/memcached": "2.2.6",
+    "@types/micro": "7.3.3",
+    "@types/nock": "10.0.3",
+    "@types/node": "8.10.62",
+    "@types/node-fetch": "2.3.2",
+    "@types/qs-middleware": "1.0.1",
+    "@types/request": "2.48.5",
+    "@types/request-promise": "4.1.46",
+    "@types/supertest": "2.0.10",
+    "@types/test-listen": "1.1.0",
+    "@types/type-is": "1.6.3",
+    "@types/uuid": "8.0.0",
+    "@types/ws": "7.2.6",
+    "apollo-fetch": "0.7.0",
+    "apollo-link": "1.2.14",
+    "apollo-link-http": "1.5.17",
+    "apollo-link-persisted-queries": "0.2.2",
+    "azure-functions-ts-essentials": "1.3.2",
+    "body-parser": "1.19.0",
+    "bunyan": "1.8.14",
+    "codecov": "3.7.2",
+    "connect": "3.7.0",
+    "deep-freeze": "0.0.1",
+    "express": "4.17.1",
+    "fastify": "1.14.6",
+    "form-data": "2.5.1",
+    "graphql": "14.7.0",
+    "graphql-subscriptions": "1.1.0",
+    "graphql-tag": "2.10.4",
+    "graphql-tools": "4.0.8",
+    "hapi": "17.8.5",
+    "ioredis": "4.17.3",
+    "jest": "25.5.4",
+    "jest-config": "25.5.4",
+    "jest-cucumber": "2.0.11",
+    "jest-junit": "10.0.0",
+    "js-sha256": "0.9.0",
+    "koa": "2.13.0",
+    "koa-router": "7.4.0",
+    "lerna": "3.22.1",
+    "log4js": "4.5.1",
+    "memcached-mock": "0.1.0",
+    "mock-req": "0.2.0",
+    "nock": "10.0.6",
+    "node-fetch": "2.3.0",
+    "prettier": "2.0.5",
+    "qs-middleware": "1.0.3",
+    "request": "2.88.2",
+    "request-promise": "4.2.5",
+    "subscriptions-transport-ws": "0.9.17",
+    "supertest": "4.0.2",
+    "test-listen": "1.1.0",
+    "ts-jest": "26.1.3",
+    "typescript": "3.9.7",
+    "winston": "3.3.3",
+    "winston-transport": "4.4.0",
+    "ws": "6.2.1"
+  },
+  "jest": {
+    "projects": [
+      "<rootDir>/packages/*"
+    ]
   }
 }
