@@ -1,22 +1,24 @@
-import { ExecutionResult } from 'graphql';
-const sha256 = require('hash.js/lib/hash/sha/256');
-
-import { CacheControlExtensionOptions } from 'apollo-cache-control';
-
-import { omit } from 'lodash';
-
-import { Request } from 'apollo-server-env';
-import { runQuery, QueryOptions } from './runQuery';
+import { Request, Headers, ValueOrPromise } from 'apollo-server-env';
 import {
   default as GraphQLOptions,
   resolveGraphqlOptions,
 } from './graphqlOptions';
 import {
+  ApolloError,
   formatApolloErrors,
   PersistedQueryNotSupportedError,
   PersistedQueryNotFoundError,
+  hasPersistedQueryError,
 } from 'apollo-server-errors';
-import { calculateCacheControlHeaders } from './caching';
+import {
+  processGraphQLRequest,
+  GraphQLRequest,
+  InvalidGraphQLRequestError,
+  GraphQLRequestContext,
+  GraphQLResponse,
+} from './requestPipeline';
+import { ApolloServerPlugin } from 'apollo-server-plugin-base';
+import { WithRequired, GraphQLExecutionResult } from 'apollo-server-types';
 
 export interface HttpQueryRequest {
   method: string;
@@ -28,13 +30,8 @@ export interface HttpQueryRequest {
   query: Record<string, any> | Array<Record<string, any>>;
   options:
     | GraphQLOptions
-    | ((...args: Array<any>) => Promise<GraphQLOptions> | GraphQLOptions);
+    | ((...args: Array<any>) => ValueOrPromise<GraphQLOptions>);
   request: Pick<Request, 'url' | 'method' | 'headers'>;
-}
-
-// The result of a curl does not appear well in the terminal, so we add an extra new line
-function prettyJSONStringify(value: any) {
-  return JSON.stringify(value) + '\n';
 }
 
 export interface ApolloServerHttpResponse {
@@ -45,6 +42,9 @@ export interface ApolloServerHttpResponse {
 }
 
 export interface HttpQueryResponse {
+  // FIXME: This isn't actually an individual GraphQL response, but the body
+  // of the HTTP response, which could contain multiple GraphQL responses
+  // when using batching.
   graphqlResponse: string;
   responseInit: ApolloServerHttpResponse;
 }
@@ -69,27 +69,45 @@ export class HttpQueryError extends Error {
 }
 
 /**
- * If optionsObject is specified, then the errors array will be formatted
+ * If options is specified, then the errors array will be formatted
  */
-function throwHttpGraphQLError<E extends Error>(
+export function throwHttpGraphQLError<E extends Error>(
   statusCode: number,
   errors: Array<E>,
-  optionsObject?: Partial<GraphQLOptions>,
+  options?: Pick<GraphQLOptions, 'debug' | 'formatError'>,
+  extensions?: GraphQLExecutionResult['extensions'],
 ): never {
+  const defaultHeaders = { 'Content-Type': 'application/json' };
+  // force no-cache on PersistedQuery errors
+  const headers = hasPersistedQueryError(errors)
+    ? {
+        ...defaultHeaders,
+        'Cache-Control': 'private, no-cache, must-revalidate',
+      }
+    : defaultHeaders;
+
+  type Result =
+   & Pick<GraphQLExecutionResult, 'extensions'>
+   & { errors: E[] | ApolloError[] }
+
+  const result: Result = {
+    errors: options
+      ? formatApolloErrors(errors, {
+          debug: options.debug,
+          formatter: options.formatError,
+        })
+      : errors,
+  };
+
+  if (extensions) {
+    result.extensions = extensions;
+  }
+
   throw new HttpQueryError(
     statusCode,
-    prettyJSONStringify({
-      errors: optionsObject
-        ? formatApolloErrors(errors, {
-            debug: optionsObject.debug,
-            formatter: optionsObject.formatError,
-          })
-        : errors,
-    }),
+    prettyJSONStringify(result),
     true,
-    {
-      'Content-Type': 'application/json',
-    },
+    headers,
   );
 }
 
@@ -97,56 +115,107 @@ export async function runHttpQuery(
   handlerArguments: Array<any>,
   request: HttpQueryRequest,
 ): Promise<HttpQueryResponse> {
-  let isGetRequest: boolean = false;
-  let optionsObject: GraphQLOptions;
+  let options: GraphQLOptions;
   const debugDefault =
     process.env.NODE_ENV !== 'production' && process.env.NODE_ENV !== 'test';
-  let cacheControl:
-    | CacheControlExtensionOptions & {
-        calculateHttpHeaders: boolean;
-        stripFormattedExtensions: boolean;
-      }
-    | undefined;
 
   try {
-    optionsObject = await resolveGraphqlOptions(
-      request.options,
-      ...handlerArguments,
-    );
+    options = await resolveGraphqlOptions(request.options, ...handlerArguments);
   } catch (e) {
     // The options can be generated asynchronously, so we don't have access to
     // the normal options provided by the user, such as: formatError,
     // debug. Therefore, we need to do some unnatural things, such
     // as use NODE_ENV to determine the debug settings
-    e.message = `Invalid options provided to ApolloServer: ${e.message}`;
-    if (!debugDefault) {
-      e.warning = `To remove the stacktrace, set the NODE_ENV environment variable to production if the options creation can fail`;
-    }
     return throwHttpGraphQLError(500, [e], { debug: debugDefault });
   }
-  if (optionsObject.debug === undefined) {
-    optionsObject.debug = debugDefault;
+  if (options.debug === undefined) {
+    options.debug = debugDefault;
   }
+
+  // FIXME: Errors thrown while resolving the context in
+  // ApolloServer#graphQLServerOptions are currently converted to
+  // a throwing function, which we invoke here to rethrow an HTTP error.
+  // When we refactor the integration between ApolloServer, the middleware and
+  // runHttpQuery, we should pass the original context function through,
+  // so we can resolve it on every GraphQL request (as opposed to once per HTTP
+  // request, which could be a batch).
+  if (typeof options.context === 'function') {
+    try {
+      (options.context as () => never)();
+    } catch (e) {
+      e.message = `Context creation failed: ${e.message}`;
+      // For errors that are not internal, such as authentication, we
+      // should provide a 400 response
+      if (
+        e.extensions &&
+        e.extensions.code &&
+        e.extensions.code !== 'INTERNAL_SERVER_ERROR'
+      ) {
+        return throwHttpGraphQLError(400, [e], options);
+      } else {
+        return throwHttpGraphQLError(500, [e], options);
+      }
+    }
+  }
+
+  const config = {
+    schema: options.schema,
+    schemaHash: options.schemaHash,
+    logger: options.logger,
+    rootValue: options.rootValue,
+    context: options.context || {},
+    validationRules: options.validationRules,
+    executor: options.executor,
+    fieldResolver: options.fieldResolver,
+
+    // FIXME: Use proper option types to ensure this
+    // The cache is guaranteed to be initialized in ApolloServer, and
+    // cacheControl defaults will also have been set if a boolean argument is
+    // passed in.
+    cache: options.cache!,
+    dataSources: options.dataSources,
+    documentStore: options.documentStore,
+
+    extensions: options.extensions,
+    persistedQueries: options.persistedQueries,
+    tracing: options.tracing,
+
+    formatError: options.formatError,
+    formatResponse: options.formatResponse,
+
+    debug: options.debug,
+
+    plugins: options.plugins || [],
+  };
+
+  return processHTTPRequest(config, request);
+}
+
+export async function processHTTPRequest<TContext>(
+  options: WithRequired<GraphQLOptions<TContext>, 'cache' | 'plugins'> & {
+    context: TContext;
+  },
+  httpRequest: HttpQueryRequest,
+): Promise<HttpQueryResponse> {
   let requestPayload;
 
-  switch (request.method) {
+  switch (httpRequest.method) {
     case 'POST':
-      if (!request.query || Object.keys(request.query).length === 0) {
+      if (!httpRequest.query || Object.keys(httpRequest.query).length === 0) {
         throw new HttpQueryError(
           500,
           'POST body missing. Did you forget use body-parser middleware?',
         );
       }
 
-      requestPayload = request.query;
+      requestPayload = httpRequest.query;
       break;
     case 'GET':
-      if (!request.query || Object.keys(request.query).length === 0) {
+      if (!httpRequest.query || Object.keys(httpRequest.query).length === 0) {
         throw new HttpQueryError(400, 'GET query missing.');
       }
 
-      isGetRequest = true;
-      requestPayload = request.query;
+      requestPayload = httpRequest.query;
       break;
 
     default:
@@ -160,277 +229,42 @@ export async function runHttpQuery(
       );
   }
 
-  let isBatch = true;
-  // TODO: do something different here if the body is an array.
-  // Throw an error if body isn't either array or object.
-  if (!Array.isArray(requestPayload)) {
-    isBatch = false;
-    requestPayload = [requestPayload];
-  }
+  // Create a local copy of `options`, based on global options, but maintaining
+  // that appropriate plugins are in place.
+  options = {
+    ...options,
+    plugins: [checkOperationPlugin, ...options.plugins],
+  };
 
-  const requests = requestPayload.map(async requestParams => {
-    try {
-      let queryString: string | undefined = requestParams.query;
-      let extensions = requestParams.extensions;
-      let persistedQueryHit = false;
-      let persistedQueryRegister = false;
-
-      if (isGetRequest && extensions) {
-        // For GET requests, we have to JSON-parse extensions. (For POST
-        // requests they get parsed as part of parsing the larger body they're
-        // inside.)
-        try {
-          extensions = JSON.parse(extensions);
-        } catch (error) {
-          throw new HttpQueryError(400, 'Extensions are invalid JSON.');
-        }
-      }
-
-      if (extensions && extensions.persistedQuery) {
-        // It looks like we've received an Apollo Persisted Query. Check if we
-        // support them. In an ideal world, we always would, however since the
-        // middleware options are created every request, it does not make sense
-        // to create a default cache here and save a referrence to use across
-        // requests
-        if (
-          !optionsObject.persistedQueries ||
-          !optionsObject.persistedQueries.cache
-        ) {
-          if (isBatch) {
-            // A batch can contain another query that returns data,
-            // so we don't error out the entire request with an HttpError
-            throw new PersistedQueryNotSupportedError();
-          }
-          // Return 200 to simplify processing: we want this to be intepreted by
-          // the client as data worth interpreting, not an error.
-          return throwHttpGraphQLError(
-            200,
-            [new PersistedQueryNotSupportedError()],
-            optionsObject,
-          );
-        } else if (extensions.persistedQuery.version !== 1) {
-          throw new HttpQueryError(400, 'Unsupported persisted query version');
-        }
-
-        const sha = extensions.persistedQuery.sha256Hash;
-
-        if (queryString === undefined) {
-          queryString =
-            (await optionsObject.persistedQueries.cache.get(`apq:${sha}`)) ||
-            undefined;
-          if (queryString) {
-            persistedQueryHit = true;
-          } else {
-            if (isBatch) {
-              // A batch can contain multiple undefined persisted queries,
-              // so we don't error out the entire request with an HttpError
-              throw new PersistedQueryNotFoundError();
-            }
-            return throwHttpGraphQLError(
-              200,
-              [new PersistedQueryNotFoundError()],
-              optionsObject,
-            );
-          }
-        } else {
-          const calculatedSha = sha256()
-            .update(queryString)
-            .digest('hex');
-          if (sha !== calculatedSha) {
-            throw new HttpQueryError(400, 'provided sha does not match query');
-          }
-          persistedQueryRegister = true;
-
-          // Do the store completely asynchronously
-          (async () => {
-            // We do not wait on the cache storage to complete
-            return (
-              optionsObject.persistedQueries &&
-              optionsObject.persistedQueries.cache.set(
-                `apq:${sha}`,
-                queryString,
-              )
-            );
-          })().catch(error => {
-            console.warn(error);
-          });
-        }
-      }
-
-      if (!queryString) {
-        throw new HttpQueryError(400, 'Must provide query string.');
-      }
-
-      if (typeof queryString !== 'string') {
-        // Check for a common error first.
-        if (queryString && (queryString as any).kind === 'Document') {
-          throw new HttpQueryError(
-            400,
-            "GraphQL queries must be strings. It looks like you're sending the " +
-              'internal graphql-js representation of a parsed query in your ' +
-              'request instead of a request in the GraphQL query language. You ' +
-              'can convert an AST to a string using the `print` function from ' +
-              '`graphql`, or use a client like `apollo-client` which converts ' +
-              'the internal representation to a string for you.',
-          );
-        }
-        throw new HttpQueryError(400, 'GraphQL queries must be strings.');
-      }
-
-      // GET operations should only be queries (not mutations). We want to throw
-      // a particular HTTP error in that case, but we don't actually parse the
-      // query until we're in runQuery, so we declare the error we want to throw
-      // here and pass it into runQuery.
-      // TODO this could/should be added as a validation rule rather than an ad hoc error
-      let nonQueryError;
-      if (isGetRequest) {
-        nonQueryError = new HttpQueryError(
-          405,
-          `GET supports only query operation`,
-          false,
-          {
-            Allow: 'POST',
-          },
-        );
-      }
-
-      const operationName = requestParams.operationName;
-
-      let variables = requestParams.variables;
-      if (typeof variables === 'string') {
-        try {
-          // XXX Really we should only do this for GET requests, but for
-          // compatibility reasons we'll keep doing this at least for now for
-          // broken clients that ship variables in a string for no good reason.
-          variables = JSON.parse(variables);
-        } catch (error) {
-          throw new HttpQueryError(400, 'Variables are invalid JSON.');
-        }
-      }
-
-      let context = optionsObject.context;
-      if (!context) {
-        context = {} as Record<string, any>;
-      } else if (typeof context === 'function') {
-        try {
-          context = await context();
-        } catch (e) {
-          e.message = `Context creation failed: ${e.message}`;
-          // For errors that are not internal, such as authentication, we
-          // should provide a 400 response
-          if (
-            e.extensions &&
-            e.extensions.code &&
-            e.extensions.code !== 'INTERNAL_SERVER_ERROR'
-          ) {
-            return throwHttpGraphQLError(400, [e], optionsObject);
-          } else {
-            return throwHttpGraphQLError(500, [e], optionsObject);
-          }
-        }
-      } else {
-        // Always clone the context if it's not a function, because that preserves
-        // having a fresh context per request.
-        context = Object.assign(
-          Object.create(Object.getPrototypeOf(context)),
-          context,
-        ) as Record<string, any>;
-      }
-
-      if (optionsObject.dataSources) {
-        const dataSources = optionsObject.dataSources() || {};
-
-        for (const dataSource of Object.values(dataSources)) {
-          if (dataSource.initialize) {
-            dataSource.initialize({ context, cache: optionsObject.cache! });
-          }
-        }
-
-        if ('dataSources' in context) {
-          throw new Error(
-            'Please use the dataSources config option instead of putting dataSources on the context yourself.',
-          );
-        }
-
-        (context as any).dataSources = dataSources;
-      }
-
-      if (optionsObject.cacheControl !== false) {
-        if (
-          typeof optionsObject.cacheControl === 'boolean' &&
-          optionsObject.cacheControl === true
-        ) {
-          // cacheControl: true means that the user needs the cache-control
-          // extensions. This means we are running the proxy, so we should not
-          // strip out the cache control extension and not add cache-control headers
-          cacheControl = {
-            stripFormattedExtensions: false,
-            calculateHttpHeaders: false,
-            defaultMaxAge: 0,
-          };
-        } else {
-          // Default behavior is to run default header calculation and return
-          // no cacheControl extensions
-          cacheControl = {
-            stripFormattedExtensions: true,
-            calculateHttpHeaders: true,
-            defaultMaxAge: 0,
-            ...optionsObject.cacheControl,
-          };
-        }
-      }
-
-      let params: QueryOptions = {
-        schema: optionsObject.schema,
-        queryString,
-        nonQueryError,
-        variables: variables,
-        context,
-        rootValue: optionsObject.rootValue,
-        operationName: operationName,
-        validationRules: optionsObject.validationRules,
-        formatError: optionsObject.formatError,
-        formatResponse: optionsObject.formatResponse,
-        fieldResolver: optionsObject.fieldResolver,
-        debug: optionsObject.debug,
-        tracing: optionsObject.tracing,
-        cacheControl: cacheControl
-          ? omit(cacheControl, [
-              'calculateHttpHeaders',
-              'stripFormattedExtensions',
-            ])
-          : false,
-        request: request.request,
-        extensions: optionsObject.extensions,
-        persistedQueryHit,
-        persistedQueryRegister,
-      };
-
-      return runQuery(params);
-    } catch (e) {
-      // Populate any HttpQueryError to our handler which should
-      // convert it to Http Error.
-      if (e.name === 'HttpQueryError') {
-        // async function wraps this in a Promise
-        throw e;
-      }
-
-      // This error will be uncaught, so we need to wrap it and treat it as an
-      // internal server error
-      return {
-        errors: formatApolloErrors([e], optionsObject),
-      };
-    }
-  }) as Array<Promise<ExecutionResult & { extensions?: Record<string, any> }>>;
-
-  let responses;
-  try {
-    responses = await Promise.all(requests);
-  } catch (e) {
-    if (e.name === 'HttpQueryError') {
-      throw e;
-    }
-    return throwHttpGraphQLError(500, [e], optionsObject);
+  function buildRequestContext(
+    request: GraphQLRequest,
+  ): GraphQLRequestContext<TContext> {
+    // FIXME: We currently shallow clone the context for every request,
+    // but that's unlikely to be what people want.
+    // We allow passing in a function for `context` to ApolloServer,
+    // but this only runs once for a batched request (because this is resolved
+    // in ApolloServer#graphQLServerOptions, before runHttpQuery is invoked).
+    // NOTE: THIS IS DUPLICATED IN ApolloServerBase.prototype.executeOperation.
+    const context = cloneObject(options.context);
+    return {
+      // While `logger` is guaranteed by internal Apollo Server usage of
+      // this `processHTTPRequest` method, this method has been publicly
+      // exported since perhaps as far back as Apollo Server 1.x.  Therefore,
+      // for compatibility reasons, we'll default to `console`.
+      logger: options.logger || console,
+      schema: options.schema,
+      schemaHash: options.schemaHash,
+      request,
+      response: {
+        http: {
+          headers: new Headers(),
+        },
+      },
+      context,
+      cache: options.cache,
+      debug: options.debug,
+      metrics: {},
+    };
   }
 
   const responseInit: ApolloServerHttpResponse = {
@@ -439,58 +273,188 @@ export async function runHttpQuery(
     },
   };
 
-  if (cacheControl) {
-    if (cacheControl.calculateHttpHeaders) {
-      const calculatedHeaders = calculateCacheControlHeaders(responses);
+  let body: string;
 
-      responseInit.headers = {
-        ...responseInit.headers,
-        ...calculatedHeaders,
-      };
-    }
+  try {
+    if (Array.isArray(requestPayload)) {
+      // We're processing a batch request
+      const requests = requestPayload.map(requestParams =>
+        parseGraphQLRequest(httpRequest.request, requestParams),
+      );
 
-    if (cacheControl.stripFormattedExtensions) {
-      responses.forEach(response => {
-        if (response.extensions) {
-          delete response.extensions.cacheControl;
-          if (Object.keys(response.extensions).length === 0) {
-            delete response.extensions;
+      const responses = await Promise.all(
+        requests.map(async request => {
+          try {
+            const requestContext = buildRequestContext(request);
+            return await processGraphQLRequest(options, requestContext);
+          } catch (error) {
+            // A batch can contain another query that returns data,
+            // so we don't error out the entire request with an HttpError
+            return {
+              errors: formatApolloErrors([error], options),
+            };
+          }
+        }),
+      );
+
+      body = prettyJSONStringify(responses.map(serializeGraphQLResponse));
+    } else {
+      // We're processing a normal request
+      const request = parseGraphQLRequest(httpRequest.request, requestPayload);
+
+      try {
+        const requestContext = buildRequestContext(request);
+
+        const response = await processGraphQLRequest(options, requestContext);
+
+        // This code is run on parse/validation errors and any other error that
+        // doesn't reach GraphQL execution
+        if (response.errors && typeof response.data === 'undefined') {
+          // don't include options, since the errors have already been formatted
+          return throwHttpGraphQLError(
+            (response.http && response.http.status) || 400,
+            response.errors as any,
+            undefined,
+            response.extensions,
+          );
+        }
+
+        if (response.http) {
+          for (const [name, value] of response.http.headers) {
+            responseInit.headers![name] = value;
           }
         }
-      });
+
+        body = prettyJSONStringify(serializeGraphQLResponse(response));
+      } catch (error) {
+        if (error instanceof InvalidGraphQLRequestError) {
+          throw new HttpQueryError(400, error.message);
+        } else if (
+          error instanceof PersistedQueryNotSupportedError ||
+          error instanceof PersistedQueryNotFoundError
+        ) {
+          return throwHttpGraphQLError(200, [error], options);
+        } else {
+          throw error;
+        }
+      }
     }
-  }
-
-  if (!isBatch) {
-    const graphqlResponse = responses[0];
-    // This code is run on parse/validation errors and any other error that
-    // doesn't reach GraphQL execution
-    if (graphqlResponse.errors && typeof graphqlResponse.data === 'undefined') {
-      // don't include optionsObject, since the errors have already been formatted
-      return throwHttpGraphQLError(400, graphqlResponse.errors as any);
+  } catch (error) {
+    if (error instanceof HttpQueryError) {
+      throw error;
     }
-    const stringified = prettyJSONStringify(graphqlResponse);
-
-    responseInit.headers!['Content-Length'] = Buffer.byteLength(
-      stringified,
-      'utf8',
-    ).toString();
-
-    return {
-      graphqlResponse: stringified,
-      responseInit,
-    };
+    return throwHttpGraphQLError(500, [error], options);
   }
-
-  const stringified = prettyJSONStringify(responses);
 
   responseInit.headers!['Content-Length'] = Buffer.byteLength(
-    stringified,
+    body,
     'utf8',
   ).toString();
 
   return {
-    graphqlResponse: stringified,
+    graphqlResponse: body,
     responseInit,
   };
+}
+
+function parseGraphQLRequest(
+  httpRequest: Pick<Request, 'url' | 'method' | 'headers'>,
+  requestParams: Record<string, any>,
+): GraphQLRequest {
+  let queryString: string | undefined = requestParams.query;
+  let extensions = requestParams.extensions;
+
+  if (typeof extensions === 'string' && extensions !== '') {
+    // For GET requests, we have to JSON-parse extensions. (For POST
+    // requests they get parsed as part of parsing the larger body they're
+    // inside.)
+    try {
+      extensions = JSON.parse(extensions);
+    } catch (error) {
+      throw new HttpQueryError(400, 'Extensions are invalid JSON.');
+    }
+  }
+
+  if (queryString && typeof queryString !== 'string') {
+    // Check for a common error first.
+    if ((queryString as any).kind === 'Document') {
+      throw new HttpQueryError(
+        400,
+        "GraphQL queries must be strings. It looks like you're sending the " +
+          'internal graphql-js representation of a parsed query in your ' +
+          'request instead of a request in the GraphQL query language. You ' +
+          'can convert an AST to a string using the `print` function from ' +
+          '`graphql`, or use a client like `apollo-client` which converts ' +
+          'the internal representation to a string for you.',
+      );
+    } else {
+      throw new HttpQueryError(400, 'GraphQL queries must be strings.');
+    }
+  }
+
+  const operationName = requestParams.operationName;
+
+  let variables = requestParams.variables;
+  if (typeof variables === 'string' && variables !== '') {
+    try {
+      // XXX Really we should only do this for GET requests, but for
+      // compatibility reasons we'll keep doing this at least for now for
+      // broken clients that ship variables in a string for no good reason.
+      variables = JSON.parse(variables);
+    } catch (error) {
+      throw new HttpQueryError(400, 'Variables are invalid JSON.');
+    }
+  }
+
+  return {
+    query: queryString,
+    operationName,
+    variables,
+    extensions,
+    http: httpRequest,
+  };
+}
+
+// GET operations should only be queries (not mutations). We want to throw
+// a particular HTTP error in that case.
+const checkOperationPlugin: ApolloServerPlugin = {
+  requestDidStart() {
+    return {
+      didResolveOperation({ request, operation }) {
+        if (!request.http) return;
+
+        if (request.http.method === 'GET' && operation.operation !== 'query') {
+          throw new HttpQueryError(
+            405,
+            `GET supports only query operation`,
+            false,
+            {
+              Allow: 'POST',
+            },
+          );
+        }
+      },
+    };
+  },
+};
+
+function serializeGraphQLResponse(
+  response: GraphQLResponse,
+): Pick<GraphQLResponse, 'errors' | 'data' | 'extensions'> {
+  // See https://github.com/facebook/graphql/pull/384 for why
+  // errors comes first.
+  return {
+    errors: response.errors,
+    data: response.data,
+    extensions: response.extensions,
+  };
+}
+
+// The result of a curl does not appear well in the terminal, so we add an extra new line
+function prettyJSONStringify(value: any) {
+  return JSON.stringify(value) + '\n';
+}
+
+export function cloneObject<T extends Object>(object: T): T {
+  return Object.assign(Object.create(Object.getPrototypeOf(object)), object);
 }
