@@ -10,16 +10,31 @@ import {
   GraphQLError,
   GraphQLNonNull,
   GraphQLScalarType,
-  introspectionQuery,
+  getIntrospectionQuery,
   BREAK,
+  DocumentNode,
+  getOperationAST,
 } from 'graphql';
 
-import request = require('supertest');
+import request from 'supertest';
 
-import { GraphQLOptions, Config } from 'apollo-server-core';
+import {
+  GraphQLOptions,
+  Config,
+  PersistedQueryOptions,
+  KeyValueCache,
+} from 'apollo-server-core';
 import gql from 'graphql-tag';
+import { ValueOrPromise } from 'apollo-server-types';
+import { GraphQLRequestListener } from "apollo-server-plugin-base";
+import { PersistedQueryNotFoundError } from "apollo-server-errors";
 
 export * from './ApolloServer';
+
+export const NODE_MAJOR_VERSION: number = parseInt(
+  process.versions.node.split('.', 1)[0],
+  10,
+);
 
 const QueryRootType = new GraphQLObjectType({
   name: 'QueryRoot',
@@ -116,7 +131,7 @@ const queryType = new GraphQLObjectType({
     },
     testContext: {
       type: GraphQLString,
-      resolve(_root, _args, context) {
+      resolve(_parent, _args, context) {
         if (context.otherField) {
           return 'unexpected';
         }
@@ -182,21 +197,25 @@ export interface CreateAppOptions {
   excludeParser?: boolean;
   graphqlOptions?:
     | GraphQLOptions
-    | { (): GraphQLOptions | Promise<GraphQLOptions> }
+    | { (): ValueOrPromise<GraphQLOptions> }
     | Config;
 }
 
 export interface CreateAppFunc {
-  (options?: CreateAppOptions): any | Promise<any>;
+  (options?: CreateAppOptions): ValueOrPromise<any>;
 }
 
 export interface DestroyAppFunc {
-  (app: any): void | Promise<void>;
+  (app: any): ValueOrPromise<void>;
 }
 
 export default (createApp: CreateAppFunc, destroyApp?: DestroyAppFunc) => {
   describe('apolloServer', () => {
     let app;
+    let didEncounterErrors: jest.Mock<
+      ReturnType<GraphQLRequestListener['didEncounterErrors']>,
+      Parameters<GraphQLRequestListener['didEncounterErrors']>
+    >;
 
     afterEach(async () => {
       if (app) {
@@ -276,22 +295,55 @@ export default (createApp: CreateAppFunc, destroyApp?: DestroyAppFunc) => {
       });
 
       it('throws error if trying to use mutation using GET request', async () => {
-        app = await createApp();
+        didEncounterErrors = jest.fn();
+        app = await createApp({
+          graphqlOptions: {
+            schema,
+            plugins: [
+              {
+                requestDidStart() {
+                  return { didEncounterErrors };
+                }
+              }
+            ]
+          }
+        });
         const query = {
           query: 'mutation test{ testMutation(echo: "ping") }',
         };
         const req = request(app)
           .get('/graphql')
           .query(query);
-        return req.then(res => {
+
+        await req.then(res => {
           expect(res.status).toEqual(405);
           expect(res.headers['allow']).toEqual('POST');
           expect(res.error.text).toMatch('GET supports only query operation');
         });
+
+        expect(didEncounterErrors).toBeCalledWith(
+          expect.objectContaining({
+            errors: expect.arrayContaining([expect.objectContaining({
+              message: 'GET supports only query operation',
+            })]),
+          }),
+        );
       });
 
       it('throws error if trying to use mutation with fragment using GET request', async () => {
-        app = await createApp();
+        didEncounterErrors = jest.fn();
+        app = await createApp({
+          graphqlOptions: {
+            schema,
+            plugins: [
+              {
+                requestDidStart() {
+                  return { didEncounterErrors };
+                }
+              }
+            ]
+          }
+        });
         const query = {
           query: `fragment PersonDetails on PersonType {
               firstName
@@ -306,11 +358,19 @@ export default (createApp: CreateAppFunc, destroyApp?: DestroyAppFunc) => {
         const req = request(app)
           .get('/graphql')
           .query(query);
-        return req.then(res => {
+        await req.then(res => {
           expect(res.status).toEqual(405);
           expect(res.headers['allow']).toEqual('POST');
           expect(res.error.text).toMatch('GET supports only query operation');
         });
+
+        expect(didEncounterErrors).toBeCalledWith(
+          expect.objectContaining({
+            errors: expect.arrayContaining([expect.objectContaining({
+              message: 'GET supports only query operation',
+            })]),
+          }),
+        );
       });
 
       it('can handle a GET request with variables', async () => {
@@ -378,7 +438,7 @@ export default (createApp: CreateAppFunc, destroyApp?: DestroyAppFunc) => {
             cacheControl: {
               defaultMaxAge: 5,
               stripFormattedExtensions: false,
-              calculateCacheControlHeaders: false,
+              calculateHttpHeaders: false,
             },
           },
         });
@@ -442,7 +502,7 @@ export default (createApp: CreateAppFunc, destroyApp?: DestroyAppFunc) => {
             },
           });
         return req.then(res => {
-          expect(res.status).toEqual(200);
+          expect(res.statusCode).toEqual(200);
           expect(res.body.errors).toBeDefined();
           expect(res.body.errors.length).toEqual(1);
           expect(res.body.errors[0].message).toEqual(
@@ -465,7 +525,7 @@ export default (createApp: CreateAppFunc, destroyApp?: DestroyAppFunc) => {
             }),
           });
         return req.then(res => {
-          expect(res.status).toEqual(200);
+          expect(res.statusCode).toEqual(200);
           expect(res.body.errors).toBeDefined();
           expect(res.body.errors.length).toEqual(1);
           expect(res.body.errors[0].message).toEqual('PersistedQueryNotFound');
@@ -565,7 +625,7 @@ export default (createApp: CreateAppFunc, destroyApp?: DestroyAppFunc) => {
         app = await createApp();
         const req = request(app)
           .post('/graphql')
-          .send({ query: introspectionQuery });
+          .send({ query: getIntrospectionQuery() });
         return req.then(res => {
           expect(res.status).toEqual(200);
           expect(res.body.data.__schema.types[0].fields[0].name).toEqual(
@@ -654,33 +714,28 @@ export default (createApp: CreateAppFunc, destroyApp?: DestroyAppFunc) => {
         });
       });
 
-      it(
-        'can handle batch requests in parallel',
-        async function() {
-          const parallels = 100;
-          const delayPerReq = 40;
+      it('can handle batch requests in parallel', async function() {
+        const parallels = 100;
+        const delayPerReq = 40;
 
-          app = await createApp();
-          const expected = Array(parallels).fill({
-            data: { testStringWithDelay: 'it works' },
-          });
-          const req = request(app)
-            .post('/graphql')
-            .send(
-              Array(parallels).fill({
-                query: `query test($delay: Int!) { testStringWithDelay(delay: $delay) }`,
-                operationName: 'test',
-                variables: { delay: delayPerReq },
-              }),
-            );
-          return req.then(res => {
-            expect(res.status).toEqual(200);
-            expect(res.body).toEqual(expected);
-          });
-        },
-        // this test will fail due to timeout if running serially.
-        3000,
-      );
+        app = await createApp();
+        const expected = Array(parallels).fill({
+          data: { testStringWithDelay: 'it works' },
+        });
+        const req = request(app)
+          .post('/graphql')
+          .send(
+            Array(parallels).fill({
+              query: `query test($delay: Int!) { testStringWithDelay(delay: $delay) }`,
+              operationName: 'test',
+              variables: { delay: delayPerReq },
+            }),
+          );
+        return req.then(res => {
+          expect(res.status).toEqual(200);
+          expect(res.body).toEqual(expected);
+        });
+      }, 3000); // this test will fail due to timeout if running serially.
 
       it('clones batch context', async () => {
         app = await createApp({
@@ -839,6 +894,40 @@ export default (createApp: CreateAppFunc, destroyApp?: DestroyAppFunc) => {
         return req.then(res => {
           expect(res.status).toEqual(200);
           expect(res.body.data.testRootValue).toEqual(expected);
+        });
+      });
+
+      it('passes the rootValue function result to the resolver', async () => {
+        const expectedQuery = 'query: it passes rootValue';
+        const expectedMutation = 'mutation: it passes rootValue';
+        app = await createApp({
+          graphqlOptions: {
+            schema,
+            rootValue: (documentNode: DocumentNode) => {
+              const op = getOperationAST(documentNode, undefined);
+              return op.operation === 'query'
+                ? expectedQuery
+                : expectedMutation;
+            },
+          },
+        });
+        const queryReq = request(app)
+          .post('/graphql')
+          .send({
+            query: 'query test{ testRootValue }',
+          });
+        return queryReq.then(res => {
+          expect(res.status).toEqual(200);
+          expect(res.body.data.testRootValue).toEqual(expectedQuery);
+        });
+        const mutationReq = request(app)
+          .post('/graphql')
+          .send({
+            query: 'mutation test{ testMutation(echo: "ping") }',
+          });
+        return mutationReq.then(res => {
+          expect(res.status).toEqual(200);
+          expect(res.body.data.testRootValue).toEqual(expectedMutation);
         });
       });
 
@@ -1028,6 +1117,90 @@ export default (createApp: CreateAppFunc, destroyApp?: DestroyAppFunc) => {
       });
     });
 
+    describe('request pipeline plugins', () => {
+      describe('lifecycle hooks', () => {
+        it('calls serverWillStart before serving a request', async () => {
+          // We'll use this eventually-assigned function to programmatically
+          // resolve the `serverWillStart` event.
+          let resolveServerWillStart: Function;
+
+          // We'll use this mocked function to determine the order in which
+          // the events we're expecting to happen actually occur and validate
+          // those expectations in various stages of this test.
+          const fn = jest.fn();
+
+          // We want this to create the app as fast as `createApp` will allow.
+          // for integrations whose `applyMiddleware` currently returns a
+          // Promise we want them to resolve at whatever eventual pace they
+          // will so we can make sure that things are happening in order.
+          const unawaitedApp = createApp({
+            graphqlOptions: {
+              schema,
+              plugins: [
+                {
+                  serverWillStart() {
+                    fn('zero');
+                    return new Promise(resolve => {
+                      resolveServerWillStart = () => {
+                        fn('one');
+                        resolve();
+                      };
+                    });
+                  },
+                },
+              ],
+            },
+          });
+
+          const delayUntil = async (check: () => boolean, expectedNumTicks) => {
+            if (check()) return expect(expectedNumTicks).toBe(0);
+            else expect(expectedNumTicks).not.toBe(0);
+            return new Promise(resolve =>
+              process.nextTick(() =>
+                delayUntil(check, expectedNumTicks - 1).then(resolve),
+              ),
+            );
+          };
+
+          // Make sure that things were called in the expected order.
+          await delayUntil(() => fn.mock.calls.length === 1, 1);
+          expect(fn.mock.calls).toEqual([['zero']]);
+          resolveServerWillStart();
+
+          // Account for the fact that `createApp` might return a Promise,
+          // and might not, depending on the integration's implementation of
+          // createApp.  This is entirely to account for the fact that
+          // non-async implementations of `applyMiddleware` leverage a
+          // middleware as the technique for yielding to `startWillStart`
+          // hooks while their `async` counterparts simply `await` those same
+          // hooks.  In a future where we make the behavior of `applyMiddleware`
+          // the same across all integrations, this should be changed to simply
+          // `await unawaitedApp`.
+          app = 'then' in unawaitedApp ? await unawaitedApp : unawaitedApp;
+
+          // Intentionally fire off the request asynchronously, without await.
+          const res = request(app)
+            .get('/graphql')
+            .query({
+              query: 'query test{ testString }',
+            })
+            .then(res => {
+              fn('two');
+              return res;
+            });
+
+          // Ensure the request has not gone through.
+          expect(fn.mock.calls).toEqual([['zero'], ['one']]);
+
+          // Now, wait for the request to finish.
+          await res;
+
+          // Finally, ensure that the order we expected was achieved.
+          expect(fn.mock.calls).toEqual([['zero'], ['one'], ['two']]);
+        });
+      });
+    });
+
     describe('Persisted Queries', () => {
       const query = '{testString}';
       const query2 = '{ testString }';
@@ -1053,25 +1226,207 @@ export default (createApp: CreateAppFunc, destroyApp?: DestroyAppFunc) => {
         },
       };
 
-      beforeEach(async () => {
+      function createMockCache(): KeyValueCache {
         const map = new Map<string, string>();
-        const cache = {
-          set: async (key, val) => {
+        return {
+          set: jest.fn(async (key, val) => {
             await map.set(key, val);
-          },
-          get: async key => map.get(key),
+          }),
+          get: jest.fn(async key => map.get(key)),
+          delete: jest.fn(async key => map.delete(key)),
         };
-        app = await createApp({
+      }
+
+      let didEncounterErrors: jest.Mock<
+        ReturnType<GraphQLRequestListener['didEncounterErrors']>,
+        Parameters<GraphQLRequestListener['didEncounterErrors']>
+      >;
+
+      let didResolveSource: jest.Mock<
+        ReturnType<GraphQLRequestListener['didResolveSource']>,
+        Parameters<GraphQLRequestListener['didResolveSource']>
+      >;
+
+      function createApqApp(apqOptions: PersistedQueryOptions = {}) {
+        return createApp({
           graphqlOptions: {
             schema,
+            plugins: [
+              {
+                requestDidStart() {
+                  return {
+                    didResolveSource,
+                    didEncounterErrors,
+                  };
+                }
+              }
+            ],
             persistedQueries: {
               cache,
+              ...apqOptions,
             },
           },
         });
+      }
+
+      let cache: KeyValueCache;
+
+      beforeEach(async () => {
+        cache = createMockCache();
+        didResolveSource = jest.fn();
+        didEncounterErrors = jest.fn();
+      });
+
+      it('when ttlSeconds is set, passes ttl to the apq cache set call', async () => {
+        app = await createApqApp({ ttl: 900 });
+
+        await request(app)
+          .post('/graphql')
+          .send({
+            extensions,
+            query,
+          });
+
+        expect(cache.set).toHaveBeenCalledWith(
+          expect.stringMatching(/^apq:/),
+          query,
+          expect.objectContaining({
+            ttl: 900,
+          }),
+        );
+        expect(didResolveSource.mock.calls[0][0])
+          .toHaveProperty('source', query);
+      });
+
+      it('when ttlSeconds is unset, ttl is not passed to apq cache',
+        async () => {
+          app = await createApqApp();
+
+          await request(app)
+            .post('/graphql')
+            .send({
+              extensions,
+              query,
+            });
+
+          expect(cache.set).toHaveBeenCalledWith(
+            expect.stringMatching(/^apq:/),
+            '{testString}',
+            expect.not.objectContaining({
+              ttl: 900,
+            }),
+          );
+          expect(didResolveSource.mock.calls[0][0])
+            .toHaveProperty('source', query);
+        }
+      );
+
+      it('errors when version is not specified', async () => {
+        app = await createApqApp();
+
+        const result = await request(app)
+          .get('/graphql')
+          .query({
+            query,
+            extensions: JSON.stringify({
+              persistedQuery: {
+                // Version intentionally omitted.
+                sha256Hash: extensions.persistedQuery.sha256Hash,
+              }
+            }),
+          });
+
+        expect(result).toMatchObject({
+          status: 400,
+          // Different integrations' response text varies in format.
+          text: expect.stringContaining('Unsupported persisted query version'),
+          req: expect.objectContaining({
+            method: 'GET',
+          }),
+        });
+
+        expect(didEncounterErrors).toBeCalledWith(
+          expect.objectContaining({
+            errors: expect.arrayContaining([expect.objectContaining({
+              message: 'Unsupported persisted query version',
+            })]),
+          }),
+        );
+      });
+
+      it('errors when version is unsupported', async () => {
+        app = await createApqApp();
+
+        const result = await request(app)
+          .get('/graphql')
+          .query({
+            query,
+            extensions: JSON.stringify({
+              persistedQuery: {
+                // Version intentionally wrong.
+                version: VERSION + 1,
+                sha256Hash: extensions.persistedQuery.sha256Hash,
+              }
+            }),
+          });
+
+        expect(result).toMatchObject({
+          status: 400,
+          // Different integrations' response text varies in format.
+          text: expect.stringContaining('Unsupported persisted query version'),
+          req: expect.objectContaining({
+            method: 'GET',
+          }),
+        });
+
+        expect(didEncounterErrors).toBeCalledWith(
+          expect.objectContaining({
+            errors: expect.arrayContaining([expect.objectContaining({
+              message: 'Unsupported persisted query version',
+            })]),
+          }),
+        );
+      });
+
+      it('errors when hash is mismatched', async () => {
+        app = await createApqApp();
+
+        const result = await request(app)
+          .get('/graphql')
+          .query({
+            query,
+            extensions: JSON.stringify({
+              persistedQuery: {
+                version: 1,
+                // Sha intentionally wrong.
+                sha256Hash: extensions.persistedQuery.sha256Hash.substr(0,5),
+              }
+            }),
+          });
+
+        expect(result).toMatchObject({
+          status: 400,
+          // Different integrations' response text varies in format.
+          text: expect.stringContaining('provided sha does not match query'),
+          req: expect.objectContaining({
+            method: 'GET',
+          }),
+        });
+
+        expect(didEncounterErrors).toBeCalledWith(
+          expect.objectContaining({
+            errors: expect.arrayContaining([expect.objectContaining({
+              message: 'provided sha does not match query',
+            })]),
+          }),
+        );
+
+        expect(didResolveSource).not.toHaveBeenCalled();
       });
 
       it('returns PersistedQueryNotFound on the first try', async () => {
+        app = await createApqApp();
+
         const result = await request(app)
           .post('/graphql')
           .send({
@@ -1084,13 +1439,38 @@ export default (createApp: CreateAppFunc, destroyApp?: DestroyAppFunc) => {
         expect(result.body.errors[0].extensions.code).toEqual(
           'PERSISTED_QUERY_NOT_FOUND',
         );
+
+        expect(didEncounterErrors).toBeCalledWith(
+          expect.objectContaining({
+            errors: expect.arrayContaining([
+              expect.any(PersistedQueryNotFoundError)
+            ]),
+          }),
+        );
+
+        expect(didResolveSource).not.toHaveBeenCalled();
       });
       it('returns result on the second try', async () => {
+        app = await createApqApp();
+
         await request(app)
           .post('/graphql')
           .send({
             extensions,
           });
+
+        // Only the first request should result in an error.
+        expect(didEncounterErrors).toHaveBeenCalledTimes(1);
+        expect(didEncounterErrors).toBeCalledWith(
+          expect.objectContaining({
+            errors: expect.arrayContaining([
+              expect.any(PersistedQueryNotFoundError)
+            ]),
+          }),
+        );
+
+        expect(didResolveSource).not.toHaveBeenCalled();
+
         const result = await request(app)
           .post('/graphql')
           .send({
@@ -1098,11 +1478,21 @@ export default (createApp: CreateAppFunc, destroyApp?: DestroyAppFunc) => {
             query,
           });
 
+        // There should be no additional errors now.  In other words, we'll
+        // re-assert that we've been called the same single time that we
+        // asserted above.
+        expect(didEncounterErrors).toHaveBeenCalledTimes(1);
+
+        expect(didResolveSource.mock.calls[0][0])
+          .toHaveProperty('source', query);
+
         expect(result.body.data).toEqual({ testString: 'it works' });
         expect(result.body.errors).toBeUndefined();
       });
 
       it('returns with batched persisted queries', async () => {
+        app = await createApqApp();
+
         const errors = await request(app)
           .post('/graphql')
           .send([
@@ -1148,11 +1538,16 @@ export default (createApp: CreateAppFunc, destroyApp?: DestroyAppFunc) => {
       });
 
       it('returns result on the persisted query', async () => {
+        app = await createApqApp();
+
         await request(app)
           .post('/graphql')
           .send({
             extensions,
           });
+
+        expect(didResolveSource).not.toHaveBeenCalled();
+
         await request(app)
           .post('/graphql')
           .send({
@@ -1165,11 +1560,16 @@ export default (createApp: CreateAppFunc, destroyApp?: DestroyAppFunc) => {
             extensions,
           });
 
+        expect(didResolveSource.mock.calls[0][0])
+          .toHaveProperty('source', query);
+
         expect(result.body.data).toEqual({ testString: 'it works' });
         expect(result.body.errors).toBeUndefined();
       });
 
       it('returns error when hash does not match', async () => {
+        app = await createApqApp();
+
         const response = await request(app)
           .post('/graphql')
           .send({
@@ -1184,9 +1584,12 @@ export default (createApp: CreateAppFunc, destroyApp?: DestroyAppFunc) => {
           });
         expect(response.status).toEqual(400);
         expect(response.error.text).toMatch(/does not match query/);
+        expect(didResolveSource).not.toHaveBeenCalled();
       });
 
       it('returns correct result using get request', async () => {
+        app = await createApqApp();
+
         await request(app)
           .post('/graphql')
           .send({
@@ -1199,6 +1602,9 @@ export default (createApp: CreateAppFunc, destroyApp?: DestroyAppFunc) => {
             extensions: JSON.stringify(extensions),
           });
         expect(result.body.data).toEqual({ testString: 'it works' });
+        expect(didResolveSource.mock.calls[0][0])
+          .toHaveProperty('source', query);
+
       });
     });
   });
