@@ -3,12 +3,10 @@ import {
   getNamedType,
   GraphQLInterfaceType,
   GraphQLObjectType,
-  GraphQLResolveInfo,
   ResponsePath,
   responsePathAsArray,
 } from 'graphql';
-
-import { GraphQLExtension } from 'graphql-extensions';
+import { ApolloServerPlugin } from "apollo-server-plugin-base";
 
 export interface CacheControlFormat {
   version: 1;
@@ -27,6 +25,10 @@ export enum CacheScope {
 
 export interface CacheControlExtensionOptions {
   defaultMaxAge?: number;
+  // FIXME: We should replace these with
+  // more appropriately named options.
+  calculateHttpHeaders?: boolean;
+  stripFormattedExtensions?: boolean;
 }
 
 declare module 'graphql/type/definition' {
@@ -38,98 +40,155 @@ declare module 'graphql/type/definition' {
   }
 }
 
-export class CacheControlExtension<TContext = any>
-  implements GraphQLExtension<TContext> {
-  private defaultMaxAge: number;
-
-  constructor(options: CacheControlExtensionOptions = {}) {
-    this.defaultMaxAge = options.defaultMaxAge || 0;
-  }
-
-  private hints: Map<ResponsePath, CacheHint> = new Map();
-
-  willResolveField(
-    _source: any,
-    _args: { [argName: string]: any },
-    _context: TContext,
-    info: GraphQLResolveInfo,
-  ) {
-    let hint: CacheHint = {};
-
-    // If this field's resolver returns an object or interface, look for hints
-    // on that return type.
-    const targetType = getNamedType(info.returnType);
-    if (
-      targetType instanceof GraphQLObjectType ||
-      targetType instanceof GraphQLInterfaceType
-    ) {
-      if (targetType.astNode) {
-        hint = mergeHints(
-          hint,
-          cacheHintFromDirectives(targetType.astNode.directives),
-        );
-      }
-    }
-
-    // If this field is a field on an object, look for hints on the field
-    // itself, taking precedence over previously calculated hints.
-    const parentType = info.parentType;
-    if (parentType instanceof GraphQLObjectType) {
-      const fieldDef = parentType.getFields()[info.fieldName];
-      if (fieldDef.astNode) {
-        hint = mergeHints(
-          hint,
-          cacheHintFromDirectives(fieldDef.astNode.directives),
-        );
-      }
-    }
-
-    // If this resolver returns an object and we haven't seen an explicit maxAge
-    // hint, set the maxAge to 0 (uncached) or the default if specified in the
-    // constructor.  (Non-object fields by default are assumed to inherit their
-    // cacheability from their parents.)
-    if (
-      (targetType instanceof GraphQLObjectType ||
-        targetType instanceof GraphQLInterfaceType) &&
-      hint.maxAge === undefined
-    ) {
-      hint.maxAge = this.defaultMaxAge;
-    }
-
-    if (hint.maxAge !== undefined || hint.scope !== undefined) {
-      this.addHint(info.path, hint);
-    }
-
-    info.cacheControl = {
-      setCacheHint: (hint: CacheHint) => {
-        this.addHint(info.path, hint);
-      },
-      cacheHint: hint,
-    };
-  }
-
-  addHint(path: ResponsePath, hint: CacheHint) {
-    const existingCacheHint = this.hints.get(path);
-    if (existingCacheHint) {
-      this.hints.set(path, mergeHints(existingCacheHint, hint));
-    } else {
-      this.hints.set(path, hint);
-    }
-  }
-
-  format(): [string, CacheControlFormat] {
-    return [
-      'cacheControl',
-      {
-        version: 1,
-        hints: Array.from(this.hints).map(([path, hint]) => ({
-          path: [...responsePathAsArray(path)],
-          ...hint,
-        })),
-      },
-    ];
+declare module 'apollo-server-types' {
+  interface GraphQLRequestContext<TContext> {
+    // Not readonly: plugins can set it.
+    overallCachePolicy?: Required<CacheHint> | undefined;
   }
 }
+
+type MapResponsePathHints = Map<ResponsePath, CacheHint>;
+
+export const plugin = (
+  options: CacheControlExtensionOptions = Object.create(null),
+): ApolloServerPlugin => ({
+  requestDidStart(requestContext) {
+    const defaultMaxAge: number = options.defaultMaxAge || 0;
+    const hints: MapResponsePathHints = new Map();
+
+
+    function setOverallCachePolicyWhenUnset() {
+      if (!requestContext.overallCachePolicy) {
+        requestContext.overallCachePolicy = computeOverallCachePolicy(hints);
+      }
+    }
+
+    return {
+      executionDidStart: () => ({
+        executionDidEnd: () => setOverallCachePolicyWhenUnset(),
+        willResolveField({ info }) {
+          let hint: CacheHint = {};
+
+          // If this field's resolver returns an object or interface, look for
+          // hints on that return type.
+          const targetType = getNamedType(info.returnType);
+          if (
+            targetType instanceof GraphQLObjectType ||
+            targetType instanceof GraphQLInterfaceType
+          ) {
+            if (targetType.astNode) {
+              hint = mergeHints(
+                hint,
+                cacheHintFromDirectives(targetType.astNode.directives),
+              );
+            }
+          }
+
+          // Look for hints on the field itself (on its parent type), taking
+          // precedence over previously calculated hints.
+          const fieldDef = info.parentType.getFields()[info.fieldName];
+          if (fieldDef.astNode) {
+            hint = mergeHints(
+              hint,
+              cacheHintFromDirectives(fieldDef.astNode.directives),
+            );
+          }
+
+          // If this resolver returns an object or is a root field and we haven't
+          // seen an explicit maxAge hint, set the maxAge to 0 (uncached) or the
+          // default if specified in the constructor. (Non-object fields by
+          // default are assumed to inherit their cacheability from their parents.
+          // But on the other hand, while root non-object fields can get explicit
+          // hints from their definition on the Query/Mutation object, if that
+          // doesn't exist then there's no parent field that would assign the
+          // default maxAge, so we do it here.)
+          if (
+            (targetType instanceof GraphQLObjectType ||
+              targetType instanceof GraphQLInterfaceType ||
+              !info.path.prev) &&
+            hint.maxAge === undefined
+          ) {
+            hint.maxAge = defaultMaxAge;
+          }
+
+          if (hint.maxAge !== undefined || hint.scope !== undefined) {
+            addHint(hints, info.path, hint);
+          }
+
+          info.cacheControl = {
+            setCacheHint: (hint: CacheHint) => {
+              addHint(hints, info.path, hint);
+            },
+            cacheHint: hint,
+          };
+        },
+      }),
+
+      responseForOperation() {
+        // We are not supplying an answer, we are only setting the cache
+        // policy if it's not set! Therefore, we return null.
+        setOverallCachePolicyWhenUnset();
+        return null;
+      },
+
+      willSendResponse(requestContext) {
+        const {
+          response,
+          overallCachePolicy: overallCachePolicyOverride,
+        } = requestContext;
+
+        // If there are any errors, we don't consider this cacheable.
+        if (response.errors) {
+          return;
+        }
+
+        // Use the override by default, but if it's not overridden, set our
+        // own computation onto the `requestContext` for other plugins to read.
+        const overallCachePolicy =
+          overallCachePolicyOverride ||
+          (requestContext.overallCachePolicy =
+            computeOverallCachePolicy(hints));
+
+        if (
+          overallCachePolicy &&
+          options.calculateHttpHeaders &&
+          response.http
+        ) {
+          response.http.headers.set(
+            'Cache-Control',
+            `max-age=${
+              overallCachePolicy.maxAge
+            }, ${overallCachePolicy.scope.toLowerCase()}`,
+          );
+        }
+
+        // We should have to explicitly ask to leave the formatted extension in,
+        // or pass the old-school `cacheControl: true` (as interpreted by
+        // apollo-server-core/ApolloServer), in order to include the
+        // engineproxy-aimed extensions. Specifically, we want users of
+        // apollo-server-plugin-response-cache to be able to specify
+        // `cacheControl: {defaultMaxAge: 600}` without accidentally turning on
+        // the extension formatting.
+        if (options.stripFormattedExtensions !== false) return;
+
+        const extensions =
+          response.extensions || (response.extensions = Object.create(null));
+
+        if (typeof extensions.cacheControl !== 'undefined') {
+          throw new Error("The cacheControl information already existed.");
+        }
+
+        extensions.cacheControl = {
+          version: 1,
+          hints: Array.from(hints).map(([path, hint]) => ({
+            path: [...responsePathAsArray(path)],
+            ...hint,
+          })),
+        };
+      }
+    }
+  }
+});
 
 function cacheHintFromDirectives(
   directives: ReadonlyArray<DirectiveNode> | undefined,
@@ -178,3 +237,45 @@ function mergeHints(
     scope: otherHint.scope || hint.scope,
   };
 }
+
+function computeOverallCachePolicy(
+  hints: MapResponsePathHints,
+): Required<CacheHint> | undefined {
+  let lowestMaxAge: number | undefined = undefined;
+  let scope: CacheScope = CacheScope.Public;
+
+  for (const hint of hints.values()) {
+    if (hint.maxAge !== undefined) {
+      lowestMaxAge =
+        lowestMaxAge !== undefined
+          ? Math.min(lowestMaxAge, hint.maxAge)
+          : hint.maxAge;
+    }
+    if (hint.scope === CacheScope.Private) {
+      scope = CacheScope.Private;
+    }
+  }
+
+  // If maxAge is 0, then we consider it uncacheable so it doesn't matter what
+  // the scope was.
+  return lowestMaxAge
+    ? {
+        maxAge: lowestMaxAge,
+        scope,
+      }
+    : undefined;
+}
+
+function addHint(hints: MapResponsePathHints, path: ResponsePath, hint: CacheHint) {
+  const existingCacheHint = hints.get(path);
+  if (existingCacheHint) {
+    hints.set(path, mergeHints(existingCacheHint, hint));
+  } else {
+    hints.set(path, hint);
+  }
+}
+
+export const __testing__ = {
+  addHint,
+  computeOverallCachePolicy,
+};
