@@ -6,5 +6,5 @@ export interface DataSourceConfig<TContext> {
 }
 
 export abstract class DataSource<TContext = any> {
-  initialize?(config: DataSourceConfig<TContext>): void;
+  initialize?(config: DataSourceConfig<TContext>): void | Promise<void>;
 }
