@@ -1,10 +1,2 @@
-export {
-  hideLiterals,
-  dropUnusedDefinitions,
-  sortAST,
-  removeAliases,
-  printWithReducedWhitespace,
-  defaultSignature,
-} from './signature';
-
 export { EngineReportingOptions, EngineReportingAgent } from './agent';
+export { default as federatedPlugin } from './federatedPlugin';
