@@ -1,34 +1,53 @@
-import { KeyValueCache } from 'apollo-server-caching';
-import * as Memcached from 'memcached';
+import {
+  TestableKeyValueCache,
+  KeyValueCacheSetOptions,
+} from 'apollo-server-caching';
+import Memcached from 'memcached';
 import { promisify } from 'util';
 
-export class MemcachedCache implements KeyValueCache {
-  readonly client;
-  readonly defaultSetOptions = {
+export class MemcachedCache implements TestableKeyValueCache {
+  // FIXME: Replace any with proper promisified type
+  readonly client: any;
+  readonly defaultSetOptions: KeyValueCacheSetOptions = {
     ttl: 300,
   };
 
   constructor(serverLocation: Memcached.Location, options?: Memcached.options) {
-    this.client = new Memcached(serverLocation, options);
+    const client = new Memcached(serverLocation, options);
     // promisify client calls for convenience
-    this.client.get = promisify(this.client.get).bind(this.client);
-    this.client.set = promisify(this.client.set).bind(this.client);
-    this.client.flush = promisify(this.client.flush).bind(this.client);
+    client.del = promisify(client.del).bind(client);
+    client.get = promisify(client.get).bind(client);
+    client.set = promisify(client.set).bind(client);
+    client.flush = promisify(client.flush).bind(client);
+
+    this.client = client;
   }
 
   async set(
     key: string,
-    data: string,
-    options?: { ttl?: number },
+    value: string,
+    options?: KeyValueCacheSetOptions,
   ): Promise<void> {
     const { ttl } = Object.assign({}, this.defaultSetOptions, options);
-    await this.client.set(key, data, ttl);
+    if (typeof ttl === 'number') {
+      await this.client.set(key, value, ttl);
+    } else {
+      // In Memcached, zero indicates no specific expiration time.  Of course,
+      // it may be purged from the cache for other reasons as deemed necessary.
+      await this.client.set(key, value, 0);
+    }
   }
 
   async get(key: string): Promise<string | undefined> {
     return await this.client.get(key);
   }
 
+  async delete(key: string): Promise<boolean> {
+    return await this.client.del(key);
+  }
+
+  // Drops all data from Memcached. This should only be used by test suites ---
+  // production code should never drop all data from an end user cache.
   async flush(): Promise<void> {
     await this.client.flush();
   }
