@@ -1,2 +1,7 @@
-export { KeyValueCache } from './KeyValueCache';
+export {
+  KeyValueCache,
+  TestableKeyValueCache,
+  KeyValueCacheSetOptions,
+} from './KeyValueCache';
 export { InMemoryLRUCache } from './InMemoryLRUCache';
+export { PrefixingKeyValueCache } from './PrefixingKeyValueCache';
