@@ -1,6 +1,5 @@
 import 'apollo-server-env';
 
-export { runQuery } from './runQuery';
 export { runHttpQuery, HttpQueryRequest, HttpQueryError } from './runHttpQuery';
 
 export {
@@ -32,6 +31,7 @@ export {
 // ApolloServer Base class
 export { ApolloServerBase } from './ApolloServer';
 export * from './types';
+export * from './requestPipelineAPI';
 
 // This currently provides the ability to have syntax highlighting as well as
 // consistency between client and server gql tags
@@ -42,6 +42,18 @@ export const gql: (
   ...substitutions: any[]
 ) => DocumentNode = gqlTag;
 
+import runtimeSupportsUploads from './utils/runtimeSupportsUploads';
 import { GraphQLScalarType } from 'graphql';
-import { GraphQLUpload as UploadScalar } from 'apollo-upload-server';
-export const GraphQLUpload = UploadScalar as GraphQLScalarType;
+export { default as processFileUploads } from './processFileUploads';
+
+// This is a conditional export intended to avoid traversing the
+// entire module tree of `graphql-upload`.  This only defined if the
+// version of Node.js is >= 8.5.0 since those are the only Node.js versions
+// which are supported by `graphql-upload@8`.  Since the source of
+// `graphql-upload` is not transpiled for older targets (in fact, it includes
+// experimental ECMAScript modules), this conditional export is necessary
+// to avoid modern ECMAScript from failing to parse by versions of Node.js
+// which don't support it (yet — eg. Node.js 6 and async/await).
+export const GraphQLUpload = runtimeSupportsUploads
+  ? (require('graphql-upload').GraphQLUpload as GraphQLScalarType)
+  : undefined;
