@@ -1,4 +1,5 @@
 import {
+  CursorShape,
   RenderPageOptions as PlaygroundRenderPageOptions,
   Theme,
 } from '@apollographql/graphql-playground-html/dist/render-playground-page';
@@ -6,16 +7,18 @@ export {
   RenderPageOptions as PlaygroundRenderPageOptions,
 } from '@apollographql/graphql-playground-html/dist/render-playground-page';
 
-// This specifies the version of GraphQL Playground that will be served
-// from graphql-playground-html, and is passed to renderPlaygroundPage
-// by the integration subclasses
-const playgroundVersion = '1.7.2';
+// This specifies the version of `graphql-playground-react` that will be served
+// from `graphql-playground-html`.  It's passed to ``graphql-playground-html`'s
+// renderPlaygroundPage` via the integration packages' playground configuration.
+const playgroundVersion = '1.7.32';
 
 // https://stackoverflow.com/a/51365037
 type RecursivePartial<T> = {
   [P in keyof T]?: T[P] extends (infer U)[]
     ? RecursivePartial<U>[]
-    : T[P] extends object ? RecursivePartial<T[P]> : T[P]
+    : T[P] extends (object | undefined)
+    ? RecursivePartial<T[P]>
+    : T[P];
 };
 
 export type PlaygroundConfig =
@@ -27,8 +30,10 @@ export const defaultPlaygroundOptions = {
   settings: {
     'general.betaUpdates': false,
     'editor.theme': 'dark' as Theme,
+    'editor.cursorShape': 'line' as CursorShape,
     'editor.reuseHeaders': true,
     'tracing.hideTracingResponse': true,
+    'queryPlan.hideQueryPlanResponse': true,
     'editor.fontSize': 14,
     'editor.fontFamily': `'Source Code Pro', 'Consolas', 'Inconsolata', 'Droid Sans Mono', 'Monaco', monospace`,
     'request.credentials': 'omit',
@@ -36,10 +41,11 @@ export const defaultPlaygroundOptions = {
 };
 
 export function createPlaygroundOptions(
-  playground: PlaygroundConfig = {},
+  playground?: PlaygroundConfig,
 ): PlaygroundRenderPageOptions | undefined {
   const isDev = process.env.NODE_ENV !== 'production';
-  const enabled: boolean = typeof playground === 'boolean' ? playground : isDev;
+  const enabled: boolean =
+    typeof playground !== 'undefined' ? !!playground : isDev;
 
   if (!enabled) {
     return undefined;
@@ -48,13 +54,19 @@ export function createPlaygroundOptions(
   const playgroundOverrides =
     typeof playground === 'boolean' ? {} : playground || {};
 
-  const playgroundOptions: PlaygroundRenderPageOptions = {
+  const settingsOverrides = playgroundOverrides.hasOwnProperty('settings')
+    ? {
+        settings: {
+          ...defaultPlaygroundOptions.settings,
+          ...playgroundOverrides.settings,
+        },
+      }
+    : { settings: undefined };
+
+  const playgroundOptions: any = {
     ...defaultPlaygroundOptions,
     ...playgroundOverrides,
-    settings: {
-      ...defaultPlaygroundOptions.settings,
-      ...playgroundOverrides.settings,
-    },
+    ...settingsOverrides,
   };
 
   return playgroundOptions;
