@@ -1,7 +1,20 @@
 import { GraphQLSchema, DocumentNode } from 'graphql';
-import { SchemaDirectiveVisitor, IResolvers, IMocks } from 'graphql-tools';
+import {
+  SchemaDirectiveVisitor,
+  IResolvers,
+  IMocks,
+  GraphQLParseOptions,
+} from 'graphql-tools';
+import {
+  ValueOrPromise,
+  GraphQLExecutor,
+  GraphQLExecutionResult,
+  GraphQLRequestContextExecutionDidStart,
+} from 'apollo-server-types';
 import { ConnectionContext } from 'subscriptions-transport-ws';
-import * as WebSocket from 'ws';
+// The types for `ws` use `export = WebSocket`, so we'll use the
+// matching `import =` to bring in its sole export.
+import WebSocket = require('ws');
 import { GraphQLExtension } from 'graphql-extensions';
 export { GraphQLExtension } from 'graphql-extensions';
 
@@ -14,13 +27,22 @@ import {
   GraphQLServerOptions as GraphQLOptions,
   PersistedQueryOptions,
 } from './graphqlOptions';
+import { CacheControlExtensionOptions } from 'apollo-cache-control';
+import { ApolloServerPlugin } from 'apollo-server-plugin-base';
+
+import { GraphQLSchemaModule } from '@apollographql/apollo-tools';
+export { GraphQLSchemaModule };
 
 export { KeyValueCache } from 'apollo-server-caching';
 
-export type Context<T = any> = T;
-export type ContextFunction<T = any> = (
-  context: Context<T>,
-) => Promise<Context<T>>;
+export type Context<T = object> = T;
+export type ContextFunction<FunctionParams = any, ProducedContext = object> = (
+  context: FunctionParams,
+) => ValueOrPromise<Context<ProducedContext>>;
+
+// A plugin can return an interface that matches `ApolloServerPlugin`, or a
+// factory function that returns `ApolloServerPlugin`.
+export type PluginDefinition = ApolloServerPlugin | (() => ApolloServerPlugin);
 
 export interface SubscriptionServerOptions {
   path: string;
@@ -33,36 +55,75 @@ export interface SubscriptionServerOptions {
   onDisconnect?: (websocket: WebSocket, context: ConnectionContext) => any;
 }
 
+type BaseConfig = Pick<
+  GraphQLOptions<Context>,
+  | 'formatError'
+  | 'debug'
+  | 'rootValue'
+  | 'validationRules'
+  | 'executor'
+  | 'formatResponse'
+  | 'fieldResolver'
+  | 'tracing'
+  | 'dataSources'
+  | 'cache'
+  | 'logger'
+>;
+
+export type Unsubscriber = () => void;
+export type SchemaChangeCallback = (schema: GraphQLSchema) => void;
+
+export type GraphQLServiceConfig = {
+  schema: GraphQLSchema;
+  executor: GraphQLExecutor;
+};
+
+/**
+ * This is a restricted view of an engine configuration which only supplies the
+ * necessary info for accessing things like cloud storage.
+ */
+export type GraphQLServiceEngineConfig = {
+  apiKeyHash: string;
+  graphId: string;
+  graphVariant?: string;
+};
+
+export interface GraphQLService {
+  load(options: {
+    engine?: GraphQLServiceEngineConfig;
+  }): Promise<GraphQLServiceConfig>;
+  onSchemaChange(callback: SchemaChangeCallback): Unsubscriber;
+  // Note: The `TContext` typing here is not conclusively behaving as we expect:
+  // https://github.com/apollographql/apollo-server/pull/3811#discussion_r387381605
+  executor<TContext>(
+    requestContext: GraphQLRequestContextExecutionDidStart<TContext>,
+  ): ValueOrPromise<GraphQLExecutionResult>;
+}
+
 // This configuration is shared between all integrations and should include
 // fields that are not specific to a single integration
-export interface Config
-  extends Pick<
-      GraphQLOptions<Context<any>>,
-      | 'formatError'
-      | 'debug'
-      | 'rootValue'
-      | 'validationRules'
-      | 'formatResponse'
-      | 'fieldResolver'
-      | 'cacheControl'
-      | 'tracing'
-      | 'dataSources'
-      | 'cache'
-    > {
-  typeDefs?: DocumentNode | Array<DocumentNode>;
-  resolvers?: IResolvers;
+export interface Config extends BaseConfig {
+  modules?: GraphQLSchemaModule[];
+  typeDefs?: DocumentNode | Array<DocumentNode> | string | Array<string>;
+  parseOptions?: GraphQLParseOptions;
+  resolvers?: IResolvers | Array<IResolvers>;
   schema?: GraphQLSchema;
   schemaDirectives?: Record<string, typeof SchemaDirectiveVisitor>;
-  context?: Context<any> | ContextFunction<any>;
+  context?: Context | ContextFunction;
   introspection?: boolean;
   mocks?: boolean | IMocks;
-  engine?: boolean | EngineReportingOptions;
+  mockEntireSchema?: boolean;
+  engine?: boolean | EngineReportingOptions<Context>;
   extensions?: Array<() => GraphQLExtension>;
+  cacheControl?: CacheControlExtensionOptions | boolean;
+  plugins?: PluginDefinition[];
   persistedQueries?: PersistedQueryOptions | false;
   subscriptions?: Partial<SubscriptionServerOptions> | string | false;
-  //https://github.com/jaydenseric/apollo-upload-server#options
+  //https://github.com/jaydenseric/graphql-upload#type-uploadoptions
   uploads?: boolean | FileUploadOptions;
   playground?: PlaygroundConfig;
+  gateway?: GraphQLService;
+  experimental_approximateDocumentStoreMiB?: number;
 }
 
 export interface FileUploadOptions {
