@@ -1,6 +1,11 @@
 import './polyfills/Object.values';
+import './polyfills/Object.entries';
 
-require('util.promisify').shim();
+import runtimeSupportsPromisify from './utils/runtimeSupportsPromisify';
+
+if (!runtimeSupportsPromisify) {
+  require('util.promisify').shim();
+}
 
 export * from './polyfills/fetch';
 export * from './polyfills/url';
