@@ -1,5 +1,5 @@
-import * as express from 'express';
-import * as corsMiddleware from 'cors';
+import express from 'express';
+import corsMiddleware from 'cors';
 import { json, OptionsJson } from 'body-parser';
 import {
   renderPlaygroundPage,
@@ -10,17 +10,27 @@ import {
   FileUploadOptions,
   ApolloServerBase,
   formatApolloErrors,
+  processFileUploads,
+  ContextFunction,
+  Context,
+  Config,
 } from 'apollo-server-core';
-import * as accepts from 'accepts';
-import * as typeis from 'type-is';
-
+import { ExecutionParams } from 'subscriptions-transport-ws';
+import accepts from 'accepts';
+import typeis from 'type-is';
 import { graphqlExpress } from './expressApollo';
-
-import { processRequest as processFileUploads } from 'apollo-upload-server';
 
 export { GraphQLOptions, GraphQLExtension } from 'apollo-server-core';
 
-export interface ServerRegistration {
+export interface GetMiddlewareOptions {
+  path?: string;
+  cors?: corsMiddleware.CorsOptions | corsMiddleware.CorsOptionsDelegate | boolean;
+  bodyParserConfig?: OptionsJson | boolean;
+  onHealthCheck?: (req: express.Request) => Promise<any>;
+  disableHealthCheck?: boolean;
+}
+
+export interface ServerRegistration extends GetMiddlewareOptions {
   // Note: You can also pass a connect.Server here. If we changed this field to
   // `express.Application | connect.Server`, it would be very hard to get the
   // app.use calls to typecheck even though they do work properly. Our
@@ -28,11 +38,6 @@ export interface ServerRegistration {
   // we suspect the only connect users left writing GraphQL apps are Meteor
   // users).
   app: express.Application;
-  path?: string;
-  cors?: corsMiddleware.CorsOptions | boolean;
-  bodyParserConfig?: OptionsJson | boolean;
-  onHealthCheck?: (req: express.Request) => Promise<any>;
-  disableHealthCheck?: boolean;
 }
 
 const fileUploadMiddleware = (
@@ -44,8 +49,11 @@ const fileUploadMiddleware = (
   next: express.NextFunction,
 ) => {
   // Note: we use typeis directly instead of via req.is for connect support.
-  if (typeis(req, ['multipart/form-data'])) {
-    processFileUploads(req, uploadsConfig)
+  if (
+    typeof processFileUploads === 'function' &&
+    typeis(req, ['multipart/form-data'])
+  ) {
+    processFileUploads(req, res, uploadsConfig)
       .then(body => {
         req.body = body;
         next();
@@ -65,7 +73,21 @@ const fileUploadMiddleware = (
   }
 };
 
+export interface ExpressContext {
+  req: express.Request;
+  res: express.Response;
+  connection?: ExecutionParams;
+}
+
+export interface ApolloServerExpressConfig extends Config {
+  context?: ContextFunction<ExpressContext, Context> | Context;
+}
+
 export class ApolloServer extends ApolloServerBase {
+  constructor(config: ApolloServerExpressConfig) {
+    super(config);
+  }
+
   // This translates the arguments from the middleware into graphQL options It
   // provides typings for the integration specific behavior, ideally this would
   // be propagated with a generic to the super class
@@ -84,19 +106,44 @@ export class ApolloServer extends ApolloServerBase {
     return true;
   }
 
-  public applyMiddleware({
-    app,
+  public applyMiddleware({ app, ...rest }: ServerRegistration) {
+    app.use(this.getMiddleware(rest));
+  }
+
+  // TODO: While `express` is not Promise-aware, this should become `async` in
+  // a major release in order to align the API with other integrations (e.g.
+  // Hapi) which must be `async`.
+  public getMiddleware({
     path,
     cors,
     bodyParserConfig,
     disableHealthCheck,
     onHealthCheck,
-  }: ServerRegistration) {
+  }: GetMiddlewareOptions = {}): express.Router {
     if (!path) path = '/graphql';
 
+    const router = express.Router();
+
+    // Despite the fact that this `applyMiddleware` function is `async` in
+    // other integrations (e.g. Hapi), currently it is not for Express (@here).
+    // That should change in a future version, but that would be a breaking
+    // change right now (see comment above this method's declaration above).
+    //
+    // That said, we do need to await the `willStart` lifecycle event which
+    // can perform work prior to serving a request.  Since Express doesn't
+    // natively support Promises yet, we'll do this via a middleware that
+    // calls `next` when the `willStart` finishes.  We'll kick off the
+    // `willStart` right away, so hopefully it'll finish before the first
+    // request comes in, but we won't call `next` on this middleware until it
+    // does. (And we'll take care to surface any errors via the `.catch`-able.)
+    const promiseWillStart = this.willStart();
+
+    router.use(path, (_req, _res, next) => {
+      promiseWillStart.then(() => next()).catch(next);
+    });
+
     if (!disableHealthCheck) {
-      // uses same path as engine proxy, but is generally useful.
-      app.use('/.well-known/apollo/server-health', (req, res) => {
+      router.use('/.well-known/apollo/server-health', (req, res) => {
         // Response follows https://tools.ietf.org/html/draft-inadarei-api-health-check-01
         res.type('application/health+json');
 
@@ -115,7 +162,7 @@ export class ApolloServer extends ApolloServerBase {
     }
 
     let uploadsMiddleware;
-    if (this.uploadsConfig) {
+    if (this.uploadsConfig && typeof processFileUploads === 'function') {
       uploadsMiddleware = fileUploadMiddleware(this.uploadsConfig, this);
     }
 
@@ -125,26 +172,26 @@ export class ApolloServer extends ApolloServerBase {
     // Note that we don't just pass all of these handlers to a single app.use call
     // for 'connect' compatibility.
     if (cors === true) {
-      app.use(path, corsMiddleware());
+      router.use(path, corsMiddleware());
     } else if (cors !== false) {
-      app.use(path, corsMiddleware(cors));
+      router.use(path, corsMiddleware(cors));
     }
 
     if (bodyParserConfig === true) {
-      app.use(path, json());
+      router.use(path, json());
     } else if (bodyParserConfig !== false) {
-      app.use(path, json(bodyParserConfig));
+      router.use(path, json(bodyParserConfig));
     }
 
     if (uploadsMiddleware) {
-      app.use(path, uploadsMiddleware);
+      router.use(path, uploadsMiddleware);
     }
 
     // Note: if you enable playground in production and expect to be able to see your
     // schema, you'll need to manually specify `introspection: true` in the
     // ApolloServer constructor; by default, the introspection query is only
     // enabled in dev.
-    app.use(path, (req, res, next) => {
+    router.use(path, (req, res, next) => {
       if (this.playgroundOptions && req.method === 'GET') {
         // perform more expensive content-type check only if necessary
         // XXX We could potentially move this logic into the GuiOptions lambda,
@@ -158,7 +205,7 @@ export class ApolloServer extends ApolloServerBase {
 
         if (prefersHTML) {
           const playgroundRenderPageOptions: PlaygroundRenderPageOptions = {
-            endpoint: path,
+            endpoint: req.originalUrl,
             subscriptionEndpoint: this.subscriptionsPath,
             ...this.playgroundOptions,
           };
@@ -169,17 +216,14 @@ export class ApolloServer extends ApolloServerBase {
           return;
         }
       }
-      return graphqlExpress(this.createGraphQLServerOptions.bind(this))(
+
+      return graphqlExpress(() => this.createGraphQLServerOptions(req, res))(
         req,
         res,
         next,
       );
     });
+
+    return router;
   }
 }
-
-export const registerServer = () => {
-  throw new Error(
-    'Please use server.applyMiddleware instead of registerServer. This warning will be removed in the next release',
-  );
-};
