@@ -1,16 +1,21 @@
-import * as express from 'express';
+import express from 'express';
 
-import * as http from 'http';
+import http from 'http';
 
-import * as request from 'request';
-import * as FormData from 'form-data';
-import * as fs from 'fs';
+import request from 'request';
+import FormData from 'form-data';
+import fs from 'fs';
 import { createApolloFetch } from 'apollo-fetch';
 
-import { gql, AuthenticationError, Config } from 'apollo-server-core';
-import { ApolloServer, ServerRegistration } from '../ApolloServer';
+import { gql, AuthenticationError } from 'apollo-server-core';
+import {
+  ApolloServer,
+  ApolloServerExpressConfig,
+  ServerRegistration,
+} from '../ApolloServer';
 
 import {
+  NODE_MAJOR_VERSION,
   testApolloServer,
   createServerInfo,
 } from 'apollo-server-integration-testsuite';
@@ -27,8 +32,6 @@ const resolvers = {
   },
 };
 
-const port = 6666;
-
 describe('apollo-server-express', () => {
   let server;
   let httpServer;
@@ -38,7 +41,7 @@ describe('apollo-server-express', () => {
       const app = express();
       server.applyMiddleware({ app });
       httpServer = await new Promise<http.Server>(resolve => {
-        const s = app.listen({ port }, () => resolve(s));
+        const s = app.listen({ port: 0 }, () => resolve(s));
       });
       return createServerInfo(server, httpServer);
     },
@@ -56,7 +59,7 @@ describe('apollo-server-express', () => {
   let httpServer: http.Server;
 
   async function createServer(
-    serverOptions: Config,
+    serverOptions: ApolloServerExpressConfig,
     options: Partial<ServerRegistration> = {},
   ) {
     server = new ApolloServer(serverOptions);
@@ -65,7 +68,7 @@ describe('apollo-server-express', () => {
     server.applyMiddleware({ ...options, app });
 
     httpServer = await new Promise<http.Server>(resolve => {
-      const l = app.listen({ port }, () => resolve(l));
+      const l = app.listen({ port: 0 }, () => resolve(l));
     });
 
     return createServerInfo(server, httpServer);
@@ -171,6 +174,7 @@ describe('apollo-server-express', () => {
               reject(error);
             } else {
               expect(body).toMatch('GraphQLPlayground');
+              expect(body).not.toMatch('settings');
               expect(response.statusCode).toEqual(200);
               resolve();
             }
@@ -179,10 +183,58 @@ describe('apollo-server-express', () => {
       });
     });
 
-    it('accepts partial GraphQL Playground Options', async () => {
+    it('renders GraphQL playground using request original url', async () => {
       const nodeEnv = process.env.NODE_ENV;
       delete process.env.NODE_ENV;
+      const samplePath = '/innerSamplePath';
 
+      const rewiredServer = new ApolloServer({
+        typeDefs,
+        resolvers,
+      });
+      const innerApp = express();
+      rewiredServer.applyMiddleware({ app: innerApp });
+
+      const outerApp = express();
+      outerApp.use(samplePath, innerApp);
+
+      const httpRewiredServer = await new Promise<http.Server>(resolve => {
+        const l = outerApp.listen({ port: 0 }, () => resolve(l));
+      });
+
+      const { url } = createServerInfo(rewiredServer, httpRewiredServer);
+
+      const paths = url.split('/');
+      const rewiredEndpoint = `${samplePath}/${paths.pop()}`;
+
+      await new Promise<http.Server>((resolve, reject) => {
+        request(
+          {
+            url: paths.join('/') + rewiredEndpoint,
+            method: 'GET',
+            headers: {
+              accept:
+                'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8',
+            },
+          },
+          (error, response, body) => {
+            process.env.NODE_ENV = nodeEnv;
+            if (error) {
+              reject(error);
+            } else {
+              expect(body).toMatch(rewiredEndpoint);
+              expect(body).not.toMatch('settings');
+              expect(response.statusCode).toEqual(200);
+              resolve();
+            }
+          },
+        );
+      });
+      await rewiredServer.stop();
+      await httpRewiredServer.close();
+    });
+
+    const playgroundPartialOptionsTest = async () => {
       const defaultQuery = 'query { foo { bar } }';
       const endpoint = '/fumanchupacabra';
       const { url } = await createServer(
@@ -220,12 +272,11 @@ describe('apollo-server-express', () => {
             },
           },
           (error, response, body) => {
-            process.env.NODE_ENV = nodeEnv;
             if (error) {
               reject(error);
             } else {
               expect(body).toMatch('GraphQLPlayground');
-              expect(body).toMatch(`"editor.theme": "light"`);
+              expect(body).toMatch(`\"editor.theme\":\"light\"`);
               expect(body).toMatch(defaultQuery);
               expect(body).toMatch(endpoint);
               expect(response.statusCode).toEqual(200);
@@ -234,7 +285,25 @@ describe('apollo-server-express', () => {
           },
         );
       });
+    };
+
+    it('accepts partial GraphQL Playground Options in production', async () => {
+      const nodeEnv = process.env.NODE_ENV;
+      process.env.NODE_ENV = 'production';
+      await playgroundPartialOptionsTest();
+      process.env.NODE_ENV = nodeEnv;
     });
+
+    it(
+      'accepts partial GraphQL Playground Options when an environment is ' +
+        'not specified',
+      async () => {
+        const nodeEnv = process.env.NODE_ENV;
+        delete process.env.NODE_ENV;
+        await playgroundPartialOptionsTest();
+        process.env.NODE_ENV = nodeEnv;
+      },
+    );
 
     it('accepts playground options as a boolean', async () => {
       const nodeEnv = process.env.NODE_ENV;
@@ -411,48 +480,47 @@ describe('apollo-server-express', () => {
         });
       });
     });
-    describe('file uploads', () => {
-      it('enabled uploads', async () => {
-        // XXX This is currently a failing test for node 10
-        const NODE_VERSION = process.version.split('.');
-        const NODE_MAJOR_VERSION = parseInt(NODE_VERSION[0].replace(/^v/, ''));
-        if (NODE_MAJOR_VERSION === 10) return;
+    // NODE: Skip Node.js 6 and 14, but only because `graphql-upload`
+    // doesn't support them on the version we use.
+    ([6, 14].includes(NODE_MAJOR_VERSION) ? describe.skip : describe)(
+      'file uploads',
+      () => {
+        it('enabled uploads', async () => {
+          const { port } = await createServer({
+            typeDefs: gql`
+              type File {
+                filename: String!
+                mimetype: String!
+                encoding: String!
+              }
 
-        const { port } = await createServer({
-          typeDefs: gql`
-            type File {
-              filename: String!
-              mimetype: String!
-              encoding: String!
-            }
+              type Query {
+                uploads: [File]
+              }
 
-            type Query {
-              uploads: [File]
-            }
-
-            type Mutation {
-              singleUpload(file: Upload!): File!
-            }
-          `,
-          resolvers: {
-            Query: {
-              uploads: () => {},
-            },
-            Mutation: {
-              singleUpload: async (_, args) => {
-                expect((await args.file).stream).toBeDefined();
-                return args.file;
+              type Mutation {
+                singleUpload(file: Upload!): File!
+              }
+            `,
+            resolvers: {
+              Query: {
+                uploads: () => {},
+              },
+              Mutation: {
+                singleUpload: async (_, args) => {
+                  expect((await args.file).stream).toBeDefined();
+                  return args.file;
+                },
               },
             },
-          },
-        });
+          });
 
-        const body = new FormData();
+          const body = new FormData();
 
-        body.append(
-          'operations',
-          JSON.stringify({
-            query: `
+          body.append(
+            'operations',
+            JSON.stringify({
+              query: `
               mutation($file: Upload!) {
                 singleUpload(file: $file) {
                   filename
@@ -461,35 +529,36 @@ describe('apollo-server-express', () => {
                 }
               }
             `,
-            variables: {
-              file: null,
-            },
-          }),
-        );
+              variables: {
+                file: null,
+              },
+            }),
+          );
 
-        body.append('map', JSON.stringify({ 1: ['variables.file'] }));
-        body.append('1', fs.createReadStream('package.json'));
+          body.append('map', JSON.stringify({ 1: ['variables.file'] }));
+          body.append('1', fs.createReadStream('package.json'));
 
-        try {
-          const resolved = await fetch(`http://localhost:${port}/graphql`, {
-            method: 'POST',
-            body: body as any,
-          });
-          const text = await resolved.text();
-          const response = JSON.parse(text);
+          try {
+            const resolved = await fetch(`http://localhost:${port}/graphql`, {
+              method: 'POST',
+              body: body as any,
+            });
+            const text = await resolved.text();
+            const response = JSON.parse(text);
 
-          expect(response.data.singleUpload).toEqual({
-            filename: 'package.json',
-            encoding: '7bit',
-            mimetype: 'application/json',
-          });
-        } catch (error) {
-          // This error began appearing randomly and seems to be a dev dependency bug.
-          // https://github.com/jaydenseric/apollo-upload-server/blob/18ecdbc7a1f8b69ad51b4affbd986400033303d4/test.js#L39-L42
-          if (error.code !== 'EPIPE') throw error;
-        }
-      });
-    });
+            expect(response.data.singleUpload).toEqual({
+              filename: 'package.json',
+              encoding: '7bit',
+              mimetype: 'application/json',
+            });
+          } catch (error) {
+            // This error began appearing randomly and seems to be a dev dependency bug.
+            // https://github.com/jaydenseric/apollo-upload-server/blob/18ecdbc7a1f8b69ad51b4affbd986400033303d4/test.js#L39-L42
+            if (error.code !== 'EPIPE') throw error;
+          }
+        });
+      },
+    );
 
     describe('errors', () => {
       it('returns thrown context error as a valid graphql result', async () => {
@@ -809,28 +878,6 @@ describe('apollo-server-express', () => {
           resolvers,
           tracing: true,
           cacheControl: true,
-        });
-
-        const apolloFetch = createApolloFetch({ uri });
-        const result = await apolloFetch({
-          query: `{ books { title author } }`,
-        });
-        expect(result.data).toEqual({ books });
-        expect(result.extensions).toBeDefined();
-        expect(result.extensions.tracing).toBeDefined();
-      });
-
-      xit('applies tracing extension with engine enabled', async () => {
-        const { url: uri } = await createServer({
-          typeDefs,
-          resolvers,
-          tracing: true,
-          engine: {
-            apiKey: 'fake',
-            maxAttempts: 0,
-            endpointUrl: 'l',
-            reportErrorFunction: () => {},
-          },
         });
 
         const apolloFetch = createApolloFetch({ uri });
