@@ -1,20 +1,19 @@
-import * as Koa from 'koa';
-import * as KoaRouter from 'koa-router';
-
-import * as http from 'http';
+import http, { Server } from 'http';
 
 import { RESTDataSource } from 'apollo-datasource-rest';
 
 import { createApolloFetch } from 'apollo-fetch';
-import { ApolloServer } from '../ApolloServer';
 
-import { createServerInfo } from 'apollo-server-integration-testsuite';
-import { gql } from '../index';
+import {
+  NODE_MAJOR_VERSION,
+  createServerInfo,
+} from 'apollo-server-integration-testsuite';
 
-const restPort = 4002;
+import { gql } from 'apollo-server-core';
 
 export class IdAPI extends RESTDataSource {
-  baseURL = `http://localhost:${restPort}/`;
+  // We will set this inside tests.
+  // baseURL = `http://localhost:${restPort}/`;
 
   async getId(id: string) {
     return this.get(`id/${id}`);
@@ -43,32 +42,49 @@ const resolvers = {
   },
 };
 
-let restCalls = 0;
-const restAPI = new Koa();
-const router = new KoaRouter();
-router.all('/id/:id', ctx => {
-  const id = ctx.params.id;
-  restCalls++;
-  ctx.set('Cache-Control', 'max-age=2000, public');
-  ctx.body = { id };
-});
+// If we're on Node.js v6, skip this test, since `koa-bodyparser` has dropped
+// support for it and there was an important update to it which we brought in
+// through https://github.com/apollographql/apollo-server/pull/3229.
+// It's worth noting that Node.js v6 has been out of Long-Term-Support status
+// for four months and is no longer recommended by the Node.js Foundation.
+(
+  NODE_MAJOR_VERSION === 6 ?
+  describe.skip :
+  describe
+)('apollo-server-koa', () => {
+  const { ApolloServer } = require('../ApolloServer');
+  const Koa = require('koa');
+  const KoaRouter = require('koa-router');
 
-router.all('/str/:id', ctx => {
-  const id = ctx.params.id;
-  restCalls++;
-  ctx.set('Cache-Control', 'max-age=2000, public');
-  ctx.body = id;
-});
+  let restCalls = 0;
+  const restAPI = new Koa();
+  const router = new KoaRouter();
+  router.all('/id/:id', ctx => {
+    const id = ctx.params.id;
+    restCalls++;
+    ctx.set('Cache-Control', 'max-age=2000, public');
+    ctx.body = { id };
+  });
 
-restAPI.use(router.routes());
-restAPI.use(router.allowedMethods());
+  router.all('/str/:id', ctx => {
+    const id = ctx.params.id;
+    restCalls++;
+    ctx.set('Cache-Control', 'max-age=2000, public');
+    ctx.body = id;
+  });
 
-describe('apollo-server-koa', () => {
-  let restServer;
+  restAPI.use(router.routes());
+  restAPI.use(router.allowedMethods());
+
+  let restServer: Server;
+  let restUrl: string;
 
   beforeAll(async () => {
-    await new Promise(resolve => {
-      restServer = restAPI.listen(restPort, resolve);
+    restUrl = await new Promise(resolve => {
+      restServer = restAPI.listen(0, () => {
+        const { port } = restServer.address();
+        resolve(`http://localhost:${port}`);
+      });
     });
   });
 
@@ -76,7 +92,7 @@ describe('apollo-server-koa', () => {
     await restServer.close();
   });
 
-  let server: ApolloServer;
+  let server: import('../ApolloServer').ApolloServer;
   let httpServer: http.Server;
 
   beforeEach(() => {
@@ -93,14 +109,16 @@ describe('apollo-server-koa', () => {
       typeDefs,
       resolvers,
       dataSources: () => ({
-        id: new IdAPI(),
+        id: new class extends IdAPI {
+          baseURL = restUrl;
+        },
       }),
     });
     const app = new Koa();
 
     server.applyMiddleware({ app });
     httpServer = await new Promise<http.Server>(resolve => {
-      const s = app.listen({ port: 7778 }, () => resolve(s));
+      const s = app.listen({ port: 0 }, () => resolve(s));
     });
     const { url: uri } = createServerInfo(server, httpServer);
 
@@ -123,14 +141,16 @@ describe('apollo-server-koa', () => {
       typeDefs,
       resolvers,
       dataSources: () => ({
-        id: new IdAPI(),
+        id: new class extends IdAPI {
+          baseURL = restUrl;
+        },
       }),
     });
     const app = new Koa();
 
     server.applyMiddleware({ app });
     httpServer = await new Promise<http.Server>(resolve => {
-      const s = app.listen({ port: 7779 }, () => resolve(s));
+      const s = app.listen({ port: 0 }, () => resolve(s));
     });
     const { url: uri } = createServerInfo(server, httpServer);
 
