@@ -1,4 +1,4 @@
-import * as request from 'request';
+import request from 'request';
 import { createApolloFetch } from 'apollo-fetch';
 
 import { gql, ApolloServer } from '../index';
@@ -24,6 +24,55 @@ describe('apollo-server', () => {
     it('accepts typeDefs and mocks', () => {
       expect(() => new ApolloServer({ typeDefs, mocks: true })).not.toThrow;
     });
+
+    // These tests are duplicates of ones in apollo-server-integration-testsuite
+    // We don't actually expect Jest to do much here, the purpose of these
+    // tests is to make sure our typings are correct, and to trigger a
+    // compile error if they aren't
+    describe('context field', () => {
+      describe('as a function', () => {
+        it('can accept and return `req`', () => {
+          expect(
+            new ApolloServer({
+              typeDefs,
+              resolvers,
+              context: ({ req }) => ({ req }),
+            }),
+          ).not.toThrow;
+        });
+
+        it('can accept nothing and return an empty object', () => {
+          expect(
+            new ApolloServer({
+              typeDefs,
+              resolvers,
+              context: () => ({}),
+            }),
+          ).not.toThrow;
+        });
+      });
+    });
+    describe('as an object', () => {
+      it('can be an empty object', () => {
+        expect(
+          new ApolloServer({
+            typeDefs,
+            resolvers,
+            context: {},
+          }),
+        ).not.toThrow;
+      });
+
+      it('can contain arbitrary values', () => {
+        expect(
+          new ApolloServer({
+            typeDefs,
+            resolvers,
+            context: { value: 'arbitrary' },
+          }),
+        ).not.toThrow;
+      });
+    });
   });
 
   describe('without registerServer', () => {
@@ -38,7 +87,7 @@ describe('apollo-server', () => {
         resolvers,
       });
 
-      const { url: uri } = await server.listen();
+      const { url: uri } = await server.listen({ port: 0 });
       const apolloFetch = createApolloFetch({ uri });
       const result = await apolloFetch({ query: '{hello}' });
 
@@ -55,7 +104,7 @@ describe('apollo-server', () => {
         resolvers,
       });
 
-      const { url } = await server.listen();
+      const { url } = await server.listen({ port: 0 });
       return new Promise((resolve, reject) => {
         request(
           {
@@ -86,7 +135,7 @@ describe('apollo-server', () => {
         resolvers,
       });
 
-      const { url: uri } = await server.listen();
+      const { url: uri } = await server.listen({ port: 0 });
 
       const apolloFetch = createApolloFetch({ uri }).useAfter(
         (response, next) => {
@@ -106,7 +155,7 @@ describe('apollo-server', () => {
         cors: { origin: 'localhost' },
       });
 
-      const { url: uri } = await server.listen();
+      const { url: uri } = await server.listen({ port: 0 });
 
       const apolloFetch = createApolloFetch({ uri }).useAfter(
         (response, next) => {
@@ -125,7 +174,7 @@ describe('apollo-server', () => {
         resolvers,
       });
 
-      const { port } = await server.listen();
+      const { port } = await server.listen({ port: 0 });
       return new Promise((resolve, reject) => {
         request(
           {
