@@ -2,14 +2,13 @@
 // an express app for you instead of applyMiddleware (which you might not even
 // use with express). The dependency is unused otherwise, so don't worry if
 // you're not using express or your version doesn't quite match up.
-import * as express from 'express';
-import * as http from 'http';
-import * as net from 'net';
+import express from 'express';
+import http from 'http';
 import {
   ApolloServer as ApolloServerBase,
   CorsOptions,
+  ApolloServerExpressConfig,
 } from 'apollo-server-express';
-import { Config } from 'apollo-server-core';
 
 export * from './exports';
 
@@ -24,12 +23,19 @@ export interface ServerInfo {
 }
 
 export class ApolloServer extends ApolloServerBase {
-  private httpServer: http.Server;
+  private httpServer?: http.Server;
   private cors?: CorsOptions | boolean;
+  private onHealthCheck?: (req: express.Request) => Promise<any>;
 
-  constructor(config: Config & { cors?: CorsOptions | boolean }) {
+  constructor(
+    config: ApolloServerExpressConfig & {
+      cors?: CorsOptions | boolean;
+      onHealthCheck?: (req: express.Request) => Promise<any>;
+    },
+  ) {
     super(config);
     this.cors = config && config.cors;
+    this.onHealthCheck = config && config.onHealthCheck;
   }
 
   private createServerInfo(
@@ -37,7 +43,16 @@ export class ApolloServer extends ApolloServerBase {
     subscriptionsPath?: string,
   ): ServerInfo {
     const serverInfo: any = {
-      ...(server.address() as net.AddressInfo),
+      // TODO: Once we bump to `@types/node@10` or higher, we can replace cast
+      // with the `net.AddressInfo` type, rather than this custom interface.
+      // Unfortunately, prior to the 10.x types, this type existed on `dgram`,
+      // but not on `net`, and in later types, the `server.address()` signature
+      // can also be a string.
+      ...(server.address() as {
+        address: string;
+        family: string;
+        port: number;
+      }),
       server,
       subscriptionsPath,
     };
@@ -82,11 +97,14 @@ export class ApolloServer extends ApolloServerBase {
     // object, so we have to create it.
     const app = express();
 
+    app.disable('x-powered-by');
+
     // provide generous values for the getting started experience
     super.applyMiddleware({
       app,
       path: '/',
       bodyParserConfig: { limit: '50mb' },
+      onHealthCheck: this.onHealthCheck,
       cors:
         typeof this.cors !== 'undefined'
           ? this.cors
@@ -95,27 +113,29 @@ export class ApolloServer extends ApolloServerBase {
             },
     });
 
-    this.httpServer = http.createServer(app);
+    const httpServer = http.createServer(app);
+    this.httpServer = httpServer;
 
     if (this.subscriptionServerOptions) {
-      this.installSubscriptionHandlers(this.httpServer);
+      this.installSubscriptionHandlers(httpServer);
     }
 
     await new Promise(resolve => {
-      this.httpServer.once('listening', resolve);
+      httpServer.once('listening', resolve);
       // If the user passed a callback to listen, it'll get called in addition
       // to our resolver. They won't have the ability to get the ServerInfo
       // object unless they use our Promise, though.
-      this.httpServer.listen(...(opts.length ? opts : [{ port: 4000 }]));
+      httpServer.listen(...(opts.length ? opts : [{ port: 4000 }]));
     });
 
-    return this.createServerInfo(this.httpServer, this.subscriptionsPath);
+    return this.createServerInfo(httpServer, this.subscriptionsPath);
   }
 
   public async stop() {
     if (this.httpServer) {
-      await new Promise(resolve => this.httpServer.close(resolve));
-      this.httpServer = null;
+      const httpServer = this.httpServer;
+      await new Promise(resolve => httpServer.close(resolve));
+      this.httpServer = undefined;
     }
     await super.stop();
   }
